% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGeneBMD.R
\name{allGeneBMD}
\alias{allGeneBMD}
\title{Calculate the all-gene cocnentraiton-response and BMDL}
\usage{
allGeneBMD(
  to.file = F,
  basedir = "../input/fcdata/",
  dataset = "DMEM_6hr_pilot_normal_00",
  metric = "z",
  l2fc.limit = 1.2
)
}
\arguments{
\item{to.file}{If TRUE, write the plots to a pdf file}

\item{basedir}{The base directory to find the raw fold change and chemcial data}

\item{dataset}{The name of the datset to be used}
}
\value{
nothing is returned
}
\description{
Calculate the all-gene cocnentraiton-response and BMDL
}
