% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfuns.R
\name{auc}
\alias{auc}
\title{Area Under the Curve}
\usage{
auc(tpr, fpr)
}
\arguments{
\item{tpr}{Vector of true positive rates.}

\item{fpr}{Vector of false positive rates.}
}
\value{
AUC
}
\description{
Compute AUC for an ROC curve.
}
\details{
Uses trapezoid rule numerical integration to approximate AUC. Will be more
accurate with more fine-grained inputs.
}
\examples{
auc(c(0,.5,1), c(0,.5,1))
auc(c(0,1,1), c(0,.5,1))
}
