% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_gene_counts.R
\name{baseline_gene_counts}
\alias{baseline_gene_counts}
\title{Gene the baseline gene counts for the cell atlas project}
\usage{
baseline_gene_counts(
  db = "httr_cell_atlas",
  dir = "../input/rawdata/cellatlas/"
)
}
\arguments{
\item{db}{The name of the Mongo database}

\item{dir}{The directory where the data will be stored

This functions takes files created by export_mongo_httr_well()
* httr_cell_atlas
* httr_tox21_cpp2}
}
\description{
Gene the baseline gene counts for the cell atlas project
}
