% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioplanet_builder.R
\name{bioplanet_builder}
\alias{bioplanet_builder}
\title{BioPlanet Builder}
\usage{
bioplanet_builder(
  pathfile = "../input/processed_pathway_data/bioplanet_pathway.csv",
  catfile = "../input/processed_pathway_data/bioplanet_pathway_category.csv",
  pwayout = "../input/processed_pathway_data/bioplanet_PATHWAYS.RData",
  pdataout = "../input/processed_pathway_data/PATHWAY_LIST_bioplanet.RData"
)
}
\arguments{
\item{pathfile}{File name of bioplanet_pathway.csv.}

\item{catfile}{File name of bioplanet_pathway_category.csv.}

\item{pwayout}{File name of bioplanet_PATHWAYS.RData}

\item{pdataout}{File name of}
}
\value{
No output.
}
\description{
Converts BioPlanet data into usable pathway data.
}
\details{
This function shows how BioPlanet data was converted to usable pathway files.
As BioPlanet is updated, this function will have to be updated. It requires
two downloaded .csv files with location specified by pathfile and catfile.
It saves usable pathway files with location specified by pwayout and
pdataout to disk.
}
