% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildChemicalMappingFile.R
\name{buildChemicalMappingFile}
\alias{buildChemicalMappingFile}
\title{Build the chemical mapping file for the MCF7 screen}
\usage{
buildChemicalMappingFile(dir = "../input/chemicals/")
}
\arguments{
\item{dir}{The directory from which to read all of the raw filesatalog file}

\item{dataset}{The name to give to the data set}

\item{filetype}{Either tsv or RData}
}
\value{
A file with the FCMAT1 data is written to "../input/fcdata/FCMAT0_",dataset,".RData"
}
\description{
Build the chemical mapping file for the MCF7 screen
}
