% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFCMAT1.R
\name{buildFCMAT1}
\alias{buildFCMAT1}
\title{Build the FCMAT1 data set}
\usage{
buildFCMAT1(
  dataset = "DMEM_6hr_pilot_normal_00",
  dir = "../input/httr_mcf7_pilot/meanncnt0_5-plateteffect_0-shrinkage_normal/DMEM_6/",
  filetype = "tsv"
)
}
\arguments{
\item{dataset}{The name to give to the data set}

\item{dir}{The directory from which to read all of the raw filesatalog file}

\item{filetype}{Either tsv or RData}
}
\value{
A file with the FCMAT1 data is written to "../input/fcdata/FCMAT0_",dataset,".RData"
}
\description{
Build the FCMAT1 data set
}
