% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFCMAT2.fromDB.R
\name{buildFCMAT2.fromDB}
\alias{buildFCMAT2.fromDB}
\title{Transpose and filter the fold change matrix FCMAT1 in long format into a
gene x sample format.}
\usage{
buildFCMAT2.fromDB(
  dataset = "mcf7_ph1_pe1_normal_block_123_excludePG",
  time = 24,
  media = "DMEM",
  dir = "../input/fcdata/",
  method = "gene",
  do.read = T
)
}
\arguments{
\item{dataset}{The name to give to the data set}

\item{dir}{The directory from which to read all of the raw filesatalog file}

\item{method}{Either "gene" or "probe"}

\item{do.read}{If TRUE, read in the FCMAT1 file and place in a global.}

\item{mc.cores}{The number of cores to use in reading the tsv files}

\item{chemical.file}{The required map from sample keys to chemical information}

\item{dsstox.file}{The information mapping chemicals to DSSTox IDs}
}
\value{
Global variables are created for the FC matrix (FCMAT2), the SE matrix (SEMAT2)
and the chemical dictionary (CHEM_DICT) which translates form the sample key
(sample_id_conc_time) to the individual components
}
\description{
Transpose and filter the fold change matrix FCMAT1 in long format into a
gene x sample format.
}
