% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFCMAT2.R
\name{buildFCMAT2}
\alias{buildFCMAT2}
\title{Transpose and filter the fold change matrix FCMAT1 in long format into a
gene x sample format.}
\usage{
buildFCMAT2(
  dataset = "DMEM_6hr_pilot_normal_pe_0",
  dir = "../input/fcdata/",
  method = "gene",
  do.read = T,
  chemical.file = "../input/chemicals/HTTr.Sample.Matrix.2017.04.24.xlsx"
)
}
\arguments{
\item{method}{=gene or probe_id}

\item{chems}{The CHEMS data frame with chemical information}
}
\value{
Global variables are created for the FC matrix (FCMAT2), the SE matrix (SEMAT2)
and the chemical dictionary (CHEM_DICT) which translates form the sample key
(sample_id_conc_time) to the individual components
}
\description{
Transpose and filter the fold change matrix FCMAT1 in long format into a
gene x sample format.
}
