% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSampleMap.R
\name{buildSampleMap}
\alias{buildSampleMap}
\title{Generate the sample_key x sample x DSSTox file}
\usage{
buildSampleMap(
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  dsstox.file = "../input/DSSTox/DSSTox_sample_map.xlsx",
  dir = "../input/fcdata/",
  outfile = "../input/chemicals/HTTr_pilot_sample_map.xlsx",
  do.read = F
)
}
\arguments{
\item{dataset}{Name of hte HTTr dataset}

\item{dsstox.file}{Name of the DSStox chemical file}

\item{dir}{Directory where the FCMAT1 files lives}

\item{outfile}{Name of the output file}

\item{do.read}{If TRUE, read in the input FCMAT1 file}
}
\description{
Generate the sample_key x sample x DSSTox file
}
