% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDEG.R
\name{calcDEG}
\alias{calcDEG}
\title{Calculate the relative variability of genes to get the DEGs}
\usage{
calcDEG(
  dataset = "mcf7_ph1_pe1_normal_good_pg",
  dir = "../input/fcdata/",
  do.read = T
)
}
\arguments{
\item{dataset}{The name to give to the data set}

\item{dir}{The directory from which to read all of the raw filesatalog file}

\item{do.read}{If TRUE, read in the HTTr data file}
}
\description{
Calculate the relative variability of genes to get the DEGs
}
