% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CHEM_DICT}
\alias{CHEM_DICT}
\title{Chemical Dictionary Example}
\format{A data frame with 80 rows and 7 variables: sample_key (sample id + conc),
  sample_id (unique identifier for each sample), conc (concentration), time 
  (length of the experiment in hours), casrn, name (chemical name), 
  dsstox_substance_id.}
\usage{
CHEM_DICT
}
\description{
Chemical dictionary for 10 randomly chosen chemicals from the phase 1 screen.
}
\keyword{datasets}
