% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMAPsignatureCorrelation.R
\name{CMAPsignatureCorrelation}
\alias{CMAPsignatureCorrelation}
\title{Build the correlation matrix for the CMAP signaturess}
\usage{
CMAPsignatureCorrelation(do.plot = F, to.file = F, ngene = 100, cutoff = 0.1)
}
\arguments{
\item{min.ngene}{Signatures will only be saved if the number of genes is >= this value}

\item{max.ngene}{Signatures will only be saved if the number of genes is <= this value}
}
\value{
No output.
}
\description{
Build the correlation matrix for the CMAP signaturess
}
