% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.fc.gsva.outlers.R
\name{compare.fc.gsva.outlers}
\alias{compare.fc.gsva.outlers}
\title{Compare the outliers from the FC and GSVA calculations}
\usage{
compare.fc.gsva.outlers(
  to.file = F,
  method2 = "gsva",
  dataset = "DMEM_6hr_pilot_none_pe_1",
  pathset = "PathwaySet_20191107"
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Compare the outliers from the FC and GSVA calculations
}
