% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffCalc.R
\name{compareSignatureCutoffs}
\alias{compareSignatureCutoffs}
\title{Generates a plot comparing two cutoff calculation methods}
\usage{
compareSignatureCutoffs(empirical_cutoffs, cutoffs, nsig = NULL)
}
\arguments{
\item{empirical_cutoffs}{output from the exportSignatureCutoffs function}

\item{cutoffs}{calculated cutoffs provided by cutoffCalc function}

\item{nsig}{number of signatures to compare}
}
\value{
ggplot object which is the scatterplot comparing the two methods
}
\description{
Generates a plot comparing two cutoff calculation methods
}
