% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatDESeq2Files.R
\name{concatDESeq2Files}
\alias{concatDESeq2Files}
\title{Concatenate the input DESeq2 files}
\usage{
concatDESeq2Files(
  dataset = "DMEM_6hr_screen_normal_pe_1",
 
    indir = "../input/httr_mcf7_screen/meanncnt0_5-plateteffect_0-shrinkage_normal_DMEM_6/",
  outdir = "../input/httr_mcf7_screen/"
)
}
\arguments{
\item{dataset}{The name of the HTTr dataset}

\item{indir}{The director to read from}

\item{outdir}{The directory to write to}
}
\description{
Concatenate the input DESeq2 files
}
