% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffCalcEmpirical.R
\name{cutoffCalc.inner.emperical}
\alias{cutoffCalc.inner.emperical}
\title{Inner function for the cutoff calculation}
\usage{
cutoffCalc.inner.emperical(signature, pval)
}
\arguments{
\item{signature}{The name of the signature  for which the cutoff is to be calculated}

\item{pval}{The p-value for the baseline distribution}

\item{covmat}{THe covariance matrix}
}
\value{
vector containing the signature, cutoff, sd, bmed
}
\description{
Inner function for the cutoff calculation
}
