% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffCalcEmpirical.R
\name{cutoffCalc.inner.empirical}
\alias{cutoffCalc.inner.empirical}
\title{Inner function for the cutoff calculation based on the empirical distributions}
\usage{
cutoffCalc.inner.empirical(signature, pval, SIGSCOREMAT)
}
\arguments{
\item{signature}{The name of the signature for which the cutoff is to be calculated}

\item{pval}{The p-value for the baseline distribution}

\item{SIGSCOREMAT}{The signature score matrix}
}
\value{
vector containing the signature, cutoff, sd (one standard deviation of the signature score matrix), bmed (median of the signature score matrix)
}
\description{
Inner function for the cutoff calculation based on the empirical distributions
}
