% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffCalc.R
\name{cutoffCalc}
\alias{cutoffCalc}
\title{Calculate the signature-wise cutoffs based on the analytical method
which does not break any correlations between genes
and compare with already defined signature_cr object (see exportSignatureCutoffs function)}
\usage{
cutoffCalc(
  sigcatalog,
  sigset,
  pval = 0.05,
  seed = 12345,
  nlowconc = 2,
  mc.cores = 1,
  dtxsid.exclude = NULL,
  do.cov = T,
  do.compare = F,
  to.file = F,
  sigdbgenelist,
  FCMAT2,
  CHEM_DICT
)
}
\arguments{
\item{sigcatalog}{The name of the signature catalog to use}

\item{sigset}{The signature set}

\item{pval}{The p-value for the baseline distribution}

\item{seed}{Random seed}

\item{nlowconc}{Only include the nlowconc concentrations for each chemical when calcualting the cutoff}

\item{mc.cores}{Number of cores to use when running parallel}

\item{dtxsid.exclude}{dtxsids to exclude, default NULL}

\item{do.cov}{If TRUE, calculate the covariance matrix and store in a global}

\item{do.compare}{If TRUE, compare the cutoffs with those from the original method with no gene-gene correlation}

\item{to.file}{If TRUE, and do.compare=TRUE, send a plot of the comparison to a file}

\item{sigdbgenelist}{full path to the signatureDB_genelist.RDS file}

\item{FCMAT2}{Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{CHEM_DICT}{Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}
}
\value{
calculated cutoffs dataframe
}
\description{
Calculate the signature-wise cutoffs based on the analytical method
which does not break any correlations between genes
and compare with already defined signature_cr object (see exportSignatureCutoffs function)
}
