% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffCalcEmpirical.R
\name{cutoffCalcEmpirical}
\alias{cutoffCalcEmpirical}
\title{Calculate the signature-wise cutoffs based on the empirical distributions
which does not break any correlations between genes}
\usage{
cutoffCalcEmpirical(
  pval = 0.05,
  nlowconc = 2,
  mc.cores = 1,
  dtxsid.exclude = NULL,
  signaturescoremat,
  CHEM_DICT
)
}
\arguments{
\item{pval}{The p-value for the baseline distribution}

\item{nlowconc}{Only include the lowest nlowconc concentrations for each chemical}

\item{mc.cores}{Number of cores to use when running parallel}

\item{dtxsid.exclude}{dtxsids to exclude, default NULL}

\item{signaturescoremat}{Dataframe with one row per chemical/conc/signature combination. Columns
are: sample_id, dtxsid, casrn, name, time, conc, sigset, signature, size, signature_score}

\item{CHEM_DICT}{Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}
}
\value{
signature-wise cutoffs dataframe
}
\description{
Calculate the signature-wise cutoffs based on the empirical distributions
which does not break any correlations between genes
}
