% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffPlot.R
\name{cutoffPlot}
\alias{cutoffPlot}
\title{Calculate the signature-wise cutoffs based on the analytical method
which does not break any correlations between genes}
\usage{
cutoffPlot(
  to.file = F,
  dataset = "mcf7_ph1_pe1_normal_block_123_allPG",
  sigset = "screen_large",
  method = "fc",
  pval = 0.05,
  nlowconc = 2
)
}
\arguments{
\item{to.file}{If TRUE, and do.compare=TRUE, send a plot of the comparison to a file}

\item{dataset}{Name of actual dataset to base cutoff on.}

\item{sigset}{THe signature set}

\item{method}{The scoring method, either fc or gsea}

\item{pval}{The p-value for the baseline distribution}

\item{nlowconc}{Only include the lowest nlowconc concentrations for each chemical}

\item{basedir}{Directory that holds FCMAT2 and CHEM_DICT files.}

\item{sigcatalog}{The name of the signature catalog to use}

\item{seed}{Random seed.}

\item{mc.cores}{NUmber of coresto use when running parallel}

\item{dtxsid.exclude}{dtxsids to exclude, default NULL}

\item{do.load}{If TRUE, reload the FCMAT2 matrix, signature catalog and chemical dictionary, and store in globals}

\item{do.cov}{If TRUE, calculate the covariance matrix and store in a global}

\item{do.compare}{If TRUE, compare the cutoffs with those from the original method with no gene-gene correlation}

\item{verbose}{If TRUE, write a line for each signature to show progress.}
}
\value{
No output.
}
\description{
Calculate the signature-wise cutoffs based on the analytical method
which does not break any correlations between genes
}
