% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENDOCRINE_PATHWAYS}
\alias{ENDOCRINE_PATHWAYS}
\title{Endocrine Pathways Example}
\format{A data frame with 44 rows and 8 variables: pathset, pathway, ngene,
ngene_in_httr, pathway_class, pathway_super_class, url, gene_list.}
\usage{
ENDOCRINE_PATHWAYS
}
\description{
Pathway dictionary for 44 ER/AR related pathways
}
\keyword{datasets}
