% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mongo_httr_well.R
\name{export_mongo_httr_well}
\alias{export_mongo_httr_well}
\title{Get the raw counts from the Mongo database}
\usage{
export_mongo_httr_well(
  db = "httr_cell_atlas",
  collection = "httr_well_trt",
  dir = "../input/rawdata/cellatlas/"
)
}
\arguments{
\item{db}{The name of the Mongo database}

\item{collection}{THe name of the collection to export}

\item{dir}{The directory where the data will be stored

httr_cell_atlas
httr_tox21_cpp2}
}
\description{
Get the raw counts from the Mongo database
}
