% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDSSToxSample.R
\name{exportDSSToxSample}
\alias{exportDSSToxSample}
\title{Generate the sample x DSSTox file}
\usage{
exportDSSToxSample(outfile = "../input/DSSTox/DSSTox_sample_map.xlsx")
}
\arguments{
\item{outfile}{Name of the file to be written}
}
\description{
Generate the sample x DSSTox file
}
