% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRawCounts.R
\name{extractRawCounts}
\alias{extractRawCounts}
\title{Get the raw counts from the Mongo database by plategroup}
\usage{
extractRawCounts(dir = "../input/rawdata/cellatlas/")
}
\arguments{
\item{dir}{The directory from which to read all of the raw filesatalog file}

\item{mc.cores}{The number of cores to use in reading the tsv files}

\item{method}{Either "gene" or "probe"}

\item{do.read}{If TRUE, read in the FCMAT1 file and place in a global.}

\item{chemical.file}{The required map from sample keys to chemical information}

\item{dsstox.file}{The information mapping chemicals to DSSTox IDs}
}
\value{
Global variables are created for the FC matrix (FCMAT2), the SE matrix (SEMAT2)
and the chemical dictionary (CHEM_DICT) which translates form the sample key
(sample_id_conc_time) to the individual components
}
\description{
THIS CODE SUPERCEDED by CODE IN HTTRPL
}
