% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC.vs.GSVA.plot.R
\name{FC.vs.GSVA.plot}
\alias{FC.vs.GSVA.plot}
\title{Plot the difference between the FC and GSVA analysis at hte signature level}
\usage{
FC.vs.GSVA.plot(
  to.file = F,
  dataset = "DMEM_6hr_pilot_normal_pe_0",
  pathset = "PathwaySet_20191107",
  cutoff = 0.5
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Plot the difference between the FC and GSVA analysis at hte signature level
}
