% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FCMAT2}
\alias{FCMAT2}
\title{Fold Change Matrix Example}
\format{A sample by gene matrix with 80 rows and 10,341 columns.}
\usage{
FCMAT2
}
\description{
Fold change matrix for 10 randomly selected chemicals from the phase 1 screen.
}
\keyword{datasets}
