% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMATrepchems.R
\name{FCMATrepchems}
\alias{FCMATrepchems}
\title{FCMAT for Replicate Chemicals}
\usage{
FCMATrepchems(study = "ph1", floor = 10, bygene = T)
}
\arguments{
\item{study}{Which replicate? "ph1" or "pilot"}

\item{floor}{Which flooring for input? 5 or 10}

\item{bygene}{If bygene is TRUE, output will be chemical/concentration by
gene, to be used for pathway analysis; otherwise, it will output 
chemical/concentration by probe id for probe concentration response
modeling.}
}
\value{
No output.
}
\description{
Generates fold change matrix for the pilot/phase 1 replicate experiment.
}
\details{
Converts deseq2 output to usable FCMAT2 matrices. Also builds CHEM_DICT files
as a subset of pre-existing CHEM_DICT files. Generated files are saved
directly to disk. Not intended to be run again, but rather to document how
it was done originally.
}
