% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixSuperTarget.R
\name{fixSuperTarget}
\alias{fixSuperTarget}
\title{Replace the super_target values in the signature output file
with ones from a new catalog}
\usage{
fixSuperTarget(
  do.read = T,
  dataset = "PFAS_U2OS",
  sigcatalog = "signatureDB_master_catalog 2021-05-10",
  sigset = "screen_large",
  method = "fc"
)
}
\arguments{
\item{do.read}{If TRUE, read in FCMAT2 to a global}

\item{dataset}{The L2fc matrix data set}

\item{sigcatalog}{THe name of the signature catalog file}

\item{sigset}{The name of the signature set to use}

\item{method}{The scoring method}
}
\value{

}
\description{
Replace the super_target values in the signature output file
with ones from a new catalog
}
