% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneBaseMeanDist.R
\name{geneBaseMeanDist}
\alias{geneBaseMeanDist}
\title{get the base mean distribution for each gene}
\usage{
geneBaseMeanDist(
  to.file = F,
  do.read = F,
  dataset = "DMEM_6hr_screen_normal_pe_1"
)
}
\arguments{
\item{to.file}{If TRUE, plot to a file}

\item{do.read}{If TRUE, read the input file into memory}

\item{dataset}{The name of the dataset}
}
\value{
No output.
}
\description{
get the base mean distribution for each gene
}
