% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneConcRespPlotWrapper.R
\name{geneConcRespPlotWrapper}
\alias{geneConcRespPlotWrapper}
\title{Wrapper for all of the conc-response plotting o genes}
\usage{
geneConcRespPlotWrapper(
  dataset = "heparg2d_toxcast_pfas_pe1_normal",
  mc.cores = 20,
  do.load = T,
  to.file = F,
  pval = 0.05,
  nametag = "_conthits",
  plotrange = c(1e-04, 100),
  onefile = T,
  chemfile = "../input/PFAS/Immuntox chemical evidence.xlsx"
)
}
\arguments{
\item{dataset}{Name of the data set.}

\item{mc.cores}{Number of cores to parallelize with.}

\item{do.load}{If TRUE, load the SIGNATURE_CR file, otherwiseassume that it is in memory
to.file to.file = T saves the output to a file; otherwise it's returned.}

\item{pval}{Desired cutoff p-value.}

\item{nametag}{Optional descriptor tag to attach to file outputs for
experimental/non-default runs.}

\item{plotrange}{The x-range of the plot as a vector of 2 elements, this can be changed for special cases, but defaults to 0.001 to 100}

\item{sigset}{Name of the signature set.}

\item{method}{Pathway scoring method in c("fc", "gsva", "mygsea")}

\item{bmr_scale}{bmr scaling factor. Default = 1.349}
}
\description{
Wrapper for all of the conc-response plotting o genes
}
