% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSlice.R
\name{geneSlice}
\alias{geneSlice}
\title{Look at concentration-slides of gene CR data to understand where burst starts}
\usage{
geneSlice(
  to.file = F,
  do.load = F,
  dataset = "mcf7_ph1_pe1_normal_block_123_allPG",
  celltype = "MCF7",
  cutoff = 0.9,
 
    chemfile = "../ERModel/ER_chems mcf7_ph1_pe1_normal_block_123_allPG estrogen 0.9 10.xlsx"
)
}
\arguments{
\item{to.file}{If TRUE, send the plots to a file}

\item{do.load}{If TRUE, load hte large HTTr data set into memory}

\item{dataset}{Name of the HTTr data set}

\item{celltype}{Name of the cell type}

\item{cutoff}{The minimum number of signatures hat have to be active in a super
target for the super target to be considered active. Default is 5}

\item{sigcatalog}{Name of the signature catalog to use}

\item{sigset}{Name of the signature set}

\item{method}{Scoring method}

\item{hccut}{Exclude rows in the data set with hitcall less than this value}

\item{tccut}{Exclude rows in the data set with top_over_cutoff less than this value}

\item{minconc}{Minimum concentration used in the plots}

\item{maxconc}{Maximum concentration used in the plots

After running this, run the following ...
superTargetPODplot
superTargetStats}
}
\description{
Look at concentration-slides of gene CR data to understand where burst starts
}
