% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httrFit.R
\name{httrFit}
\alias{httrFit}
\title{HTTr Fit}
\usage{
httrFit(conc, resp, cutoff, force.fit = FALSE, bidirectional = TRUE,
  verbose = FALSE, do.plot = F, fitmodels = c("cnst", "hill", "gnls",
  "poly1", "poly2", "pow", "exp2", "exp3", "exp4", "exp5"), ...)
}
\arguments{
\item{conc}{Vector of concentrations (NOT in log units).}

\item{resp}{Vector of responses.}

\item{cutoff}{Desired cutoff. If no absolute responses > cutoff and 
force.fit = F, will only fit constant model.}

\item{force.fit}{If force.fit = T, will fit all models regardless of cutoff.}

\item{bidirectional}{If bidirectional = F, will only give positive fits.}

\item{verbose}{If verbose = T, will print optimization details and aics.}

\item{do.plot}{If do.plot = T, will generate a plot comparing model curves.}

\item{fitmodels}{Vector of model names to try fitting. Missing models still
return a skeleton output filled with NAs.}

\item{...}{Other fitting parameters (deprecated).}
}
\value{
List of 11 elements. First 10 elements are the output generated by each
  fit function with their given model names. Last element is "modelnames": a 
  vector of model names so other functions can easily cycle through the output.
}
\description{
Concentration response curve fitting for HTTr.
}
\details{
All models are monotonic and equal to 0 at 0 concentration (zero background).
To add more models in the future, write a fit____ function, and add the model
name to the fitmodels and modelnames vectors.
}
\examples{
conc = c(.03,.1,.3,1,3,10,30,100)
resp = c(0,.1,0,.2,.6,.9,1.1,1)
output = httrFit(conc,resp, .8, fitmodels = c("cnst", "hill"),verbose = TRUE, 
  do.plot = TRUE)
}
