% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeFCgenes.R
\name{largeFCgenes}
\alias{largeFCgenes}
\title{Find genes that are often seen with large fold changes}
\usage{
largeFCgenes(
  to.file = F,
  basedir = "../input/fcdata/",
  dataset = "DMEM_6hr_pilot_normal_00",
  l2fc.limit = 1.2
)
}
\arguments{
\item{to.file}{If TRUE, write the plots to a pdf file}

\item{basedir}{The base directory to find the raw fold change and chemcial data}

\item{dataset}{The name of the datset to be used}
}
\value{
nothing is returned
}
\description{
Find genes that are often seen with large fold changes
}
