% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePFASFCMAT1.R
\name{mergePFASFCMAT1}
\alias{mergePFASFCMAT1}
\title{special code the merge the PFAS replacemnte data in with the earlier data fro U2OS and HepaRG}
\usage{
mergePFASFCMAT1(
  dataset = "heparg2d_toxcast_pfas_pe1_normal_v2",
 
    file1 = "httr_heparg2d_toxcast_pfas_FCmat1-meanncnt0_5-plateteffect_1-shrinkage_normal.RData",
 
    file2 = "httr_pfas_replace_heparg_FCmat1-meanncnt0_5-plateteffect_1-shrinkage_normal.RData",
  dir = "../input/fcdata/new_versions/",
  do.load = T
)
}
\arguments{
\item{dataset}{The name to give to the data set}

\item{dir}{The directory from which to read all of the raw files}

\item{do.load}{If TRUE, read the large input data file into memory}

\item{infile}{The nae of the input file}

\item{pg.filter.file}{An optional file to use in filtering out bad plate groups}
}
\value{
A file with the FCMAT1 data is written to "../input/fcdata/FCMAT1_",dataset,".RData"

* heparg2d_toxcast_pfas_pe1_normal
* u2os_toxcast_pfas_pe1_normal
}
\description{
version to start with Logan's database export
The difference between this version and the original is that there are extra columns
The function just changes one column name and writes the file to a standard name and place
}
