% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYGSEA.R
\name{MYGSEA}
\alias{MYGSEA}
\title{My Gene Set Enrichment Analysis}
\usage{
MYGSEA(
  X,
  geneSets,
  min.sz = 1,
  max.sz = Inf,
  alpha = 0.25,
  verbose = T,
  useranks = T
)
}
\arguments{
\item{X}{Transposed FCMAT2; i.e a gene by sample matrix of l2fc's including
rownames and colnames. Equivalent to expr in gsva.}

\item{geneSets}{Named list of signature definitions. Each element is a vector
of gene names. Each element name is a signature name.E quivalent to
gset.idx.list in gsva.}

\item{min.sz}{Minimum signature size (deprecated).}

\item{max.sz}{Maximum signature size (deprecated)}

\item{alpha}{Power of R to use. Higher alpha will upweight more extreme
ranks relative to middle ranks.}

\item{verbose}{verbose = T prints gene set length message.}

\item{useranks}{useranks = T uses ranks as in ssGSEA, while useranks = F
uses the bare fold changes instead.}
}
\value{
Outputs signature by sample matrix of signature scores.
}
\description{
Performs tweaked version of single sample GSEA.
}
\details{
Based on the GSVA ssGSEA code. Main changes are: NAs are now handled correctly
and rank is now centered on zero instead of beginning at one. Since signature
sizes are undercounted here due to missing values, they are
assessed more accurately in signatureScoreCoreMYGSEA and limits are enforced
after scoring.
}
\examples{
geneSets = list(signature1 = c("ABC", "DEF"), signature2 = c("ABC", "GHI"))
X = matrix(c(1:3,3:1), nrow = 3)
colnames(X) = c("Sample1", "Sample2")
rownames(X) = c("ABC", "DEF", "GHI")
MYGSEA(X,geneSets)
}
