% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{myrndwalk}
\alias{myrndwalk}
\title{GSVA:::rndwalk but able to handle sum(Ralpha) = 0}
\usage{
myrndwalk(gSetIdx, geneRanking, j, R, alpha)
}
\arguments{
\item{gSetIdx}{gSetIdx}

\item{geneRanking}{geneRanking}

\item{j}{j}

\item{R}{R}

\item{alpha}{alpha}
}
\description{
GSVA:::rndwalk but able to handle sum(Ralpha) = 0
}
