% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeFCMAT2.R
\name{normalizeFCMAT2}
\alias{normalizeFCMAT2}
\title{Normalize FCMAT2 so tht within each plate group, the lowest 2
concentrations of each gene have a mean of zero}
\usage{
normalizeFCMAT2(
  dataset = "DMEM_6hr_screen_normal_pe_1",
  dir = "../input/fcdata/",
  method = "gene",
  do.read = F
)
}
\arguments{
\item{method}{=gene or probe_id}

\item{chems}{The CHEMS data frame with chemical information}
}
\value{
Global variables are created for the FC matrix (FCMAT2), the SE matrix (SEMAT2)
and the chemical dictionary (CHEM_DICT) which translates form the sample key
(sample_id_conc_time) to the individual components
}
\description{
Normalize FCMAT2 so tht within each plate group, the lowest 2
concentrations of each gene have a mean of zero
}
