% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway_data}
\alias{pathway_data}
\title{Endocrine Pathway Data}
\format{A named list containing 44 elements. Each elements corresponds to a
  pathway and consists of a vector of gene names.}
\usage{
pathway_data
}
\description{
Pathway data example for 44 ER/AR related pathways
}
\keyword{datasets}
