% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAccumNullPlot.R
\name{pathwayAccumNullPlot}
\alias{pathwayAccumNullPlot}
\title{BMD Accumulation Plot With Nulls}
\usage{
pathwayAccumNullPlot(pathset = "bhrr", dataset = "arer",
  method = "fc", nullset = "arer_RAND125", newpval = 0.05,
  oldpval = 0.05, to.file = T, usecont = T, nametag = NULL,
  mc.cores = 1)
}
\arguments{
\item{pathset}{Name of pathset.}

\item{dataset}{Name of dataset.}

\item{method}{Name of pathway scoring method.}

\item{nullset}{Name of null dataset.}

\item{newpval}{P-value for cutoff to be used in plot.}

\item{oldpval}{P-value that nullset and dataset were originally run with.}

\item{to.file}{If to.file = T, plots to file.}

\item{usecont}{Set usecont = T for continuous hitcalls, usecont = F for discrete.
Should probably match the original hitcall type use in CR modeling.}

\item{nametag}{Set additional name descriptor that was attached to CR modeling,
if applicable.}

\item{mc.cores}{Number of cores to use when altering continuous hitcalls; has
no effect if usecont = F or newpval = oldpval.}
}
\value{
No output.
}
\description{
Creates pathway BMD accumulation plot vs. null and computes accumulation BMD.
}
\details{
Nullset and dataset should already have been run through pathwayConcResp using
the given pathset, method, and oldpval. Only generates proof of concept plots
for accumulation BMDs. There is not currently a method to extract the 
accumulation BMDs directly.
}
