% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayConcResp_pval.R
\name{pathwayConcResp_pval}
\alias{pathwayConcResp_pval}
\title{Pathway Concentration Response (P-value)}
\usage{
pathwayConcResp_pval(pathset = "bhrr", dataset = "arer",
  method = "fc", nullset = "arer_RAND1000", mc.cores = 1,
  to.file = T, do.plot = F, pval = 0.05, nametag = NULL,
  conthits = T, aicc = F, minpathsize = 10, fitmodels = c("cnst",
  "hill", "gnls", "poly1", "poly2", "pow", "exp2", "exp3", "exp4", "exp5"))
}
\arguments{
\item{pathset}{Name of the pathway set.}

\item{dataset}{Name of the data set.}

\item{method}{Name of the pathway scoring method.}

\item{nullset}{Name of the null data set.}

\item{mc.cores}{Number of cores to parallelize with.}

\item{to.file}{to.file = T saves the output to a file; otherwise it's returned.}

\item{do.plot}{do.plot = T creates concentration-response plots for every
sample/pathway combination and saves to disk.}

\item{pval}{Desired cutoff p-value.}

\item{nametag}{Optional descriptor tag to attach to file outputs for
experimental/non-default runs.}

\item{conthits}{conthits = T uses continuous hitcalls, otherwise it's discrete.}

\item{aicc}{aicc = T uses corrected AIC to choose winning method; otherwise
regular AIC.}

\item{minpathsize}{Minimum allowed pathway size. Sample/pathway combinations 
with less than this number of non-missing l2fc's will be discarded.}

\item{fitmodels}{Vector of model names to use. Probably should include "cnst".}
}
\value{
If to.file = T, nothing. If to.file = F, dataframe with pathway CR
  output.
}
\description{
Performs pathway concentration response using p-value based cutoffs.
}
\details{
Null dataset and dataset should have already been scored using pathwayScore
and the given pathset and method. This function prepares pathwayScore output
for CR processing, calls pathwayConcRespCore_pval, formats the output,
saves it to disk, then calls plotouter for CR plots, if desired. If
conthits = T and nametag = NULL, the nametag "conthits" is automatically added
to the output file.
}
