% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayConcRespPlot.R
\name{pathwayConcRespPlot}
\alias{pathwayConcRespPlot}
\title{Pathway Concentration Response Plot}
\usage{
pathwayConcRespPlot(row)
}
\arguments{
\item{row}{Named list containing:
\itemize{
  \item conc - conc string separated by |'s
  \item resp - response string separated by |'s
  \item method - scoring method determines plot bounds
  \item proper_name - chemical name for plot title
  \item cutoff - noise cutoff
  \item bmr - baseline median response; level at which bmd is calculated
  \item er - fitted error term for plotting error bars
  \item a, tp, b, ga, p, la, q - other model parameters for fit curve
  \item fit_method - curve fit method
  \item bmd10, bmdl, bmdu - bmd, bmd lower bound, and bmd upper bound
  \item ac50, acc - curve value at 50% of top, curve value at cutoff
  \item top - curve top
  \item time, pathway, pathway_class, pathway_size - other identifiers
}
Other elements are ignored.}
}
\value{
No output.
}
\description{
Plots a concentration response curve for one sample/pathway combination.
}
\details{
row is one row of PATHWAY_CR, the pathwayConcResp output.
}
