% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayScore.R
\name{pathwayScore}
\alias{pathwayScore}
\title{Pathway Score}
\usage{
pathwayScore(FCMAT2, CHEM_DICT, pathset = "bhrr",
  dataset = "PlateEffect", method = "10chems", mc.cores = 1,
  minpathsize = 10)
}
\arguments{
\item{FCMAT2}{Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{CHEM_DICT}{Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dsstox_substance_id.}

\item{pathset}{Name of pathway set.}

\item{dataset}{Name of data set.}

\item{method}{Name of desired scoring method.}

\item{mc.cores}{Number of cores to use.}

\item{minpathsize}{Minimum allowed pathway size BEFORE accounting for
missing values.}
}
\value{
No output.
}
\description{
Computes and saves pathway scores.
}
\details{
pathwayScore is a driver for various scoring methods. The three that are
currently available are "gsva", "mygsea", "fc", and "mygsea_norank" (a version
of mygsea that uses fold changes instead of ranks as weights). Deprecated
methods include the Fisher method and gsvae (gsva with empirical cdfs). 
Beware running out of memory on large runs with gsva, Linux, and many cores.
Pathway size is counted according to number of genes in the pathway that are 
also in the column names of FCMAT2. However, each method performs a more
rigorous size count internally that accounts for missing values and adds this
to the output. This minpathsize is enforced when running pathwayConcResp_pval.
}
