% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayScoreCoreFC.R
\name{pathwayScoreCoreFC}
\alias{pathwayScoreCoreFC}
\title{Pathway Score Core - FC}
\usage{
pathwayScoreCoreFC(fcdata, pathset, dataset, chem_dict, pathway_data)
}
\arguments{
\item{fcdata}{Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{pathset}{Name of pathway set.}

\item{dataset}{Name of data set.}

\item{chem_dict}{Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dsstox_substance_id.}

\item{pathway_data}{Named ist of gene name vectors. Each element is one 
pathway, defined by the genes it contains.}
}
\value{
Dataframe with one row per chemical/conc/pathway combination. Columns
  are: sample_id, dsstox_substance_id, casrn, name, time, conc, pathset, 
  pathway, size (pathway size accounting for missing values), mean_fc_scaled_in,
  mean_fc_scaled_out, pathway_score.
}
\description{
Computes fold change pathway scores.
}
\details{
This fast implementation of fold change pathway scores uses matrix
multiplication. The score is simply: mean(fold change of genes in pathway) -
mean(fold change of genes outside pathway).
}
