% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayScoreCoreMYGSEA.R
\name{pathwayScoreCoreMYGSEA}
\alias{pathwayScoreCoreMYGSEA}
\title{Pathway Score Core - MYGSEA}
\usage{
pathwayScoreCoreMYGSEA(sk.list, method = "mygsea", pathset = "bhrr",
  dataset, fcmat, chem_dict, pathway_data, mc.cores = 1,
  normalization = T, useranks = T)
}
\arguments{
\item{sk.list}{Sample keys to use; should correspond to fcmat rownames.}

\item{method}{Method name to use in file output. "mygsea" or "mygsea_norank"}

\item{pathset}{Name of pathway set.}

\item{dataset}{Name of data set.}

\item{fcmat}{Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{chem_dict}{Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dsstox_substance_id.}

\item{pathway_data}{Named ist of gene name vectors. Each element is one 
pathway, defined by the genes it contains.}

\item{mc.cores}{Number of cores to use. Parallelization is performed 
by gsva itself.}

\item{normalization}{normalization = T normalizes final scores.}

\item{useranks}{useranks = T uses score ranks for weighting; otherwise,
fold changes are used for weights.}
}
\value{
No output.
}
\description{
Computes pathway scores for mygsea.
}
\details{
This function is a parallelized wrapper for MYGSEA, which does the actual
scoring. mygsea method uses ranks and normalization, while mygsea_norank
method does not use ranks or normalization. Normalization divides final 
scores by difference between max and min score. Without normalization,
scores from individual samples have no impact on each other. Final
pathscoremat is written to disk.
}
