% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_id.to.sample_id.R
\name{pg_id.to.sample_id}
\alias{pg_id.to.sample_id}
\title{get the mapping between the plate groups and the samples}
\usage{
pg_id.to.sample_id(
  do.load = F,
  dataset = "mcf7_ph1_pe1_normal_block_123_allPG",
  dir = "../input/fcdata/"
)
}
\arguments{
\item{do.load}{if T, load the initial file}

\item{dataset}{The name to give to the data set}

\item{dir}{The directory from which to read all of the raw files}
}
\value{
A mapping the sampels to the plate groups
f
}
\description{
version to start with Logan's database export
The difference between this version and the original is that there are extra columns
The function just changes one column name and writes the file to a standard name and place
}
