% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repChemPidPlot.R
\name{pidbar}
\alias{pidbar}
\title{PID Bar Plot}
\usage{
pidbar(x, ...)
}
\arguments{
\item{x}{Named matrix or vector to pass to barplot.}

\item{...}{Other options to pass to barplot.}
}
\value{
No output.
}
\description{
Specially formatted bar plot.
}
\details{
This function is a helper for repChemPidPlot. It fiddles with the margins
and renames the labels so that they fit on the plot.
}
