% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureConcResp.R
\name{plotouter}
\alias{plotouter}
\title{Plot Outer}
\usage{
plotouter(proper_name, SIGNATURE_CR, foldname, CYTOTOX)
}
\arguments{
\item{proper_name}{Chemical name to be used in file name.}

\item{SIGNATURE_CR}{Dataframe output of signatureConcResp_pval.}

\item{foldname}{Folder name for output file.}

\item{CYTOTOX}{The cytotoxicity data for all chemicals}
}
\value{
No output.
}
\description{
Calls signatureConcResp plotting function.
}
\details{
Calls signatureConcResp plotting function for one chemical and every signature.
Saves a single pdf to disk for the given chemical containing every signature
CR plot.
}
