% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotouterGene.R
\name{plotouterGene}
\alias{plotouterGene}
\title{Plot Outer}
\usage{
plotouterGene(proper_name, GENE_CR, foldname, plotrange = c(0.001, 100))
}
\arguments{
\item{proper_name}{Chemical name to be used in file name.}

\item{GENE_CR}{Dataframe output of geneConcResp_pval.}

\item{foldname}{Folder name for output file.}

\item{plotrange}{The x-range of the plot as a vector of 2 elements, this can be changed for special cases, but defaults to 0.001 to 100}
}
\value{
No output.
}
\description{
Calls signatureConcResp plotting function.
}
\details{
Calls signatureConcResp plotting function for one chemical and every signature.
Saves a single pdf to disk for the given chemical containing every signature
CR plot.
}
