% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podLaneplot.R
\name{podLaneplot}
\alias{podLaneplot}
\title{Build lane plots by chemical list and signature class, across the datasets}
\usage{
podLaneplot(to.file = F, dataset, sigset, method, plot.signature_min = F)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Build lane plots by chemical list and signature class, across the datasets
}
