% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repChemSignaturePlot.R
\name{pwaybar}
\alias{pwaybar}
\title{Pathway Bar Plot}
\usage{
pwaybar(x, ...)
}
\arguments{
\item{x}{Named matrix or vector to pass to barplot.}

\item{...}{Other options to pass to barplot.}
}
\value{
No output.
}
\description{
Specially formatted bar plot.
}
\details{
This function is a helper for repChemPathwayPlot. It fiddles with the margins
and renames the labels so that they fit on the plot.
}
