% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfuns.R
\name{R2}
\alias{R2}
\title{R Squared}
\usage{
R2(y, pred)
}
\arguments{
\item{y}{Vector of actual values.}

\item{pred}{Vector of corresponding predicted values.}
}
\value{
Coefficient of determination.
}
\description{
Calculate coefficient of determination.
}
\details{
Note that order matters: R2(x,y) != R2(y,x) in general.
}
\examples{
R2(c(1:10), c(1:10*.8))
R2(c(1:10*.8), c(1:10))
}
