% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomdata.R
\name{randomdata}
\alias{randomdata}
\title{Randomized Null Data}
\usage{
randomdata(
  nchem = 1000,
  seed = 12345,
  maxconc = 1e+06,
  nlowconc = 2,
  dtxsid.exclude = NULL,
  FCMAT2,
  CHEM_DICT,
  output_dir = "../data/",
  writing_to_disk = FALSE
)
}
\arguments{
\item{nchem}{Number of null chemicals. Number of null samples is approximately
eight times this value.}

\item{seed}{Random seed.}

\item{maxconc}{Only use concentrations less than maxconc, default 1000000}

\item{nlowconc}{If not NULL, only include the lowest nlowconc concentrations for each chemical}

\item{dtxsid.exclude}{dtxsids to exclude, default NULL}

\item{FCMAT2}{Sample by gene matrix of log2(fold change)'s. Rownames are sample keys and colnames are genes.}

\item{CHEM_DICT}{Data frame with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}

\item{output_dir}{full path to directory/file where the CHEM_DICT/FCMAT2 data frames should be saved}

\item{writing_to_disk}{boolean, TRUE if the resulting CHEM_DICT/FCMAT2 data frames are to be saved to disk}
}
\value{
list of the randomized FCMAT2 and CHEM_DICT data frame
}
\description{
Generate randomized null data based on actual data.
}
\details{
Randomization is performed by sampling the quantile function for each gene in the actual data.
The null set will have roughly the same distribution of values for each gene in the actual data.
}
