% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repChemPathwayPlot.R
\name{repChemPathwayPlot}
\alias{repChemPathwayPlot}
\title{Replicate Chemical Pathway Plot}
\usage{
repChemPathwayPlot(oldpval = 0.05, nametag = "conthits",
  method = "fc", pathset = "bhrr", mc.cores = 3)
}
\arguments{
\item{oldpval}{P-value used to generate PATHWAY_CR's.}

\item{nametag}{Optional descriptor in filename.}

\item{method}{Name of pathway scoring method used.}

\item{pathset}{Name of pathway set.}

\item{mc.cores}{Number of cores to use.}
}
\value{
No output.
}
\description{
Generates plots and statistics for replicate chemicals' pathways.
}
\details{
This function is desigend to work with runAllRepChemCR, so the dataset names
are hard-coded. This function may take some time to run. Concentration response
should have been run using continuous hitcalls.
}
