% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repChemPidPlot.R
\name{repChemPidPlot}
\alias{repChemPidPlot}
\title{Replicate Chemical PID Plot}
\usage{
repChemPidPlot(oldpval = 0.05, nametag = "conthits", mc.cores = 3)
}
\arguments{
\item{oldpval}{P-value used to generate GENE_CR's.}

\item{nametag}{Optional descriptor in filename.}

\item{mc.cores}{Number of cores to use.}
}
\value{
No output.
}
\description{
Generates plots and statistics for replicate chemicals' probe IDs.
}
\details{
This function is desigend to work with runAllRepChemPidCR, so the dataset names
are hard-coded. This function may take some time to run. Concentration response
should have been run using continuous hitcalls.
}
