% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfuns.R
\name{RMSE}
\alias{RMSE}
\title{Root-mean-square-error}
\usage{
RMSE(x, y)
}
\arguments{
\item{x}{First vector.}

\item{y}{Second vector.}
}
\value{
RMSE
}
\description{
Computes root-mean-square-error between two vectors.
}
\examples{
RMSE(1:3, c(1,3,5))
}
