% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffCalc.R
\name{rowMean}
\alias{rowMean}
\title{Utility function used by cutoffCalc.inner.fc}
\usage{
rowMean(x)
}
\arguments{
\item{x}{= matrix or data.frame}
}
\value{
vector of mean value of each row
}
\description{
Utility function used by cutoffCalc.inner.fc
}
