% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAllPathwayCR_pval.R
\name{runAllPathwayCR_pval}
\alias{runAllPathwayCR_pval}
\title{Run All Pathway Concentration Response (P-Value)}
\usage{
runAllPathwayCR_pval(basedir = "input/fcdata/", dataset = "arer",
  pathset = "bhrr", method = "fc", minpathsize = 10, conthits = T,
  nullset = "arer_RAND125", do.plot = T, pval = 0.05,
  mc.cores = c(39, 39), fitmodels = c("cnst", "hill", "gnls", "poly1",
  "poly2", "pow", "exp2", "exp3", "exp4", "exp5"))
}
\arguments{
\item{basedir}{Folder that stores FCMAT2 and CHEM_DICT files.}

\item{dataset}{Name of data set.}

\item{pathset}{Name of pathway set.}

\item{method}{Name of pathway scoring method.}

\item{minpathsize}{Minimum pathway size.}

\item{conthits}{conthits = T uses continous hitcall; conthits = F uses discrete
hitcalls.}

\item{nullset}{Name of null dataset. Set nullset = NULL to skip CR.}

\item{do.plot}{do.plot = T generates a CR plot for every sample/pathway
combination.}

\item{pval}{P-value to use for noise estimation.}

\item{mc.cores}{Vector with two values: number of cores to use for pathway
scoring and number of cores to use for CR. CR can usually handle the maximum
number, but gsva scoring might require a smaller number to avoid memory
overflow.}

\item{fitmodels}{Vector of model names to run conc/resp with. "cnst" should 
always be chosen.}
}
\value{
No output.
}
\description{
Driver for pathway scoring and concentration response (CR).
}
\details{
CR requires pathway scores to have already been computed for a nullset.
randomdata() can generate a nullset, and this function can compute pathway
scores for it by setting dataset = nullset and nullset = NULL. Pathway
scores are written to disk in output/pathway_score_summary/. CR results
are written to disk in output/pathway/conc_resp_summary/.
}
