% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAllRepChemCR.R
\name{runAllRepChemCR}
\alias{runAllRepChemCR}
\title{Run All Replicate Chemical Concentration Response}
\usage{
runAllRepChemCR(
  basedir = "../input/fcdata/",
  pathset = "bhrr",
  method = "fc",
  minpathsize = 10,
  do.plot = F,
  pval = 0.05,
  mc.cores = c(39, 39),
  conthits = T,
  nchem = 125
)
}
\arguments{
\item{basedir}{Folder that the FCMAT2's are stored in.}

\item{pathset}{Name of signature set.}

\item{method}{Name of signature scoring method.}

\item{minpathsize}{Minimum signature size.}

\item{do.plot}{do.plot = T generates plots for every chemical/signature/replicate
combination. Adds a significant amount to the runtime.}

\item{pval}{P-value to use for noise estimation.}

\item{mc.cores}{Vector with two values: number of cores to use for signature
scoring and number of cores to use for CR. CR can usually handle the maximum
number, but gsva scoring might require a smaller number to avoid memory
overflow.}

\item{conthits}{conthits = T uses continuous hitcalls. Continuous hitcalls are
a prerequisitie for using repChemPathwayPlot().}

\item{nchem}{Number of null chemicals to use. The number of null samples is
approximately eight times this value, so nchem = 125 generates ~1000 null
samples.}
}
\value{
No output.
}
\description{
Runs signature scoring and concentration response for replicate chemicals.
}
\details{
This function has hard-coded dataset names for the replicates. For each
replicate, it computes signature scores, generates a null dataset, runs
signature scores for the null dataset, and then runs concentration-response
on the actual data. Pathway scores and CR are written to disk.
}
