% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAllRepChemPidCR.R
\name{runAllRepChemPidCR}
\alias{runAllRepChemPidCR}
\title{Run All Replicate Chemical PID Concentration Response}
\usage{
runAllRepChemPidCR(pval = 0.05, mc.cores = 39, conthits = T)
}
\arguments{
\item{pval}{P-value to use for noise estimation. Noise is estimated using two 
lowest concentrations.}

\item{mc.cores}{Number of cores to use for CR.}

\item{conthits}{conthits = T uses continuous hitcalls. Continuous hitcalls are
a prerequisitie for using repChemPidPlot().}
}
\value{
No output.
}
\description{
Runs probe ID concentration response for replicate chemicals.
}
\details{
This function has hard-coded dataset names for the replicates. For each
replicate, it runs concentration-response directly on the probe ID's. The
result is written to disk.
}
