% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureAccumNullPlot.R
\name{signatureAccumNullPlot}
\alias{signatureAccumNullPlot}
\title{BMD Accumulation Plot With Nulls}
\usage{
signatureAccumNullPlot(
  pathset = "PathwaySet_20191107",
  dataset = "DMEM_6hr_pilot_normal_pe_0",
  method = "fc",
  nullset = "DMEM_6hr_pilot_normal_pe_0_RAND1000",
  newpval = 0.05,
  oldpval = 0.05,
  to.file = F,
  usecont = T,
  nametag = "conthits",
  mc.cores = 1,
  hit.threshold = 0.5,
  verbose = F
)
}
\arguments{
\item{pathset}{Name of pathset.}

\item{dataset}{Name of dataset.}

\item{method}{Name of signature scoring method.}

\item{nullset}{Name of null dataset.}

\item{newpval}{P-value for cutoff to be used in plot.}

\item{oldpval}{P-value that nullset and dataset were originally run with.}

\item{to.file}{If to.file = T, plots to file.}

\item{usecont}{Set usecont = T for continuous hitcalls, usecont = F for discrete.
Should probably match the original hitcall type use in CR modeling.}

\item{nametag}{Set additional name descriptor that was attached to CR modeling,
if applicable.}

\item{mc.cores}{Number of cores to use when altering continuous hitcalls; has
no effect if usecont = F or newpval = oldpval.}
}
\value{
No output.
}
\description{
Creates signature BMD accumulation plot vs. null and computes accumulation BMD.
}
\details{
Nullset and dataset should already have been run through signatureConcResp using
the given pathset, method, and oldpval. Only generates proof of concept plots
for accumulation BMDs. There is not currently a method to extract the
accumulation BMDs directly.
}
