% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBmdeDeseq.R
\name{signatureBmdeDeseq}
\alias{signatureBmdeDeseq}
\title{Compare the signature PODs between BMDExpress and DESeq2 / TCPL}
\usage{
signatureBmdeDeseq(
  to.file = F,
  dataset = "DMEM_6hr_pilot_normal_pe_0",
  pathset = "PathwaySet_20191107",
  method = "fc"
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Compare the signature PODs between BMDExpress and DESeq2 / TCPL
}
