% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuildCMAP.R
\name{signatureBuildCMAP}
\alias{signatureBuildCMAP}
\title{Build the standard input file for the CMAP signatures}
\usage{
signatureBuildCMAP()
}
\value{
No output.
}
\description{
Build the standard input file for the CMAP signatures
}
