% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuildDisGeNET.R
\name{signatureBuildDisGeNET}
\alias{signatureBuildDisGeNET}
\title{Build the standard input file for the Ryan signatures}
\usage{
signatureBuildDisGeNET()
}
\value{
No output.
}
\description{
Build the standard input file for the Ryan signatures
}
