% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuilder_bhrr.R
\name{signatureBuilder_bhrr}
\alias{signatureBuilder_bhrr}
\title{bhrr Pathway Builder}
\usage{
signatureBuilder_bhrr()
}
\value{
No output.
}
\description{
Builds bhrr pathset based on msigdb, bioplanet, and ryan pathsets.
}
\details{
Shows how the bhrr pathset was built from pre-existing pathsets.
}
