% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuilderRandom.R
\name{signatureBuilderRandom}
\alias{signatureBuilderRandom}
\title{Add the random gene sets to the signature files}
\usage{
signatureBuilderRandom(nrandom = 1000, mc.cores = 1)
}
\arguments{
\item{nrandom}{Number of random gene sets}

\item{mc.cores}{The number of cores to use in parallel}
}
\value{
No output.
}
\description{
Add the random gene sets to the signature files
}
