% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuildMsigDB.R
\name{signatureBuildMsigDB}
\alias{signatureBuildMsigDB}
\title{Build the standard input file for the MSigDB signatures}
\usage{
signatureBuildMsigDB()
}
\value{
No output.
}
\description{
Build the standard input file for the MSigDB signatures
}
