% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureBuildRyan.R
\name{signatureBuildRyan}
\alias{signatureBuildRyan}
\title{Build the standard input file for the Ryan signatures}
\usage{
signatureBuildRyan()
}
\value{
No output.
}
\description{
Build the standard input file for the Ryan signatures
}
