% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureCatalogLoader.R
\name{signatureCatalogLoader}
\alias{signatureCatalogLoader}
\title{Read in signature catalog}
\usage{
signatureCatalogLoader(
  sigset = "wgcna",
  sigcatalog = "../inst/extdata/signatureDB_master_catalog_2022-05-16.xlsx",
  sigdbgenelist = "../inst/extdata/signatureDB_genelists.RDS"
)
}
\arguments{
\item{sigset}{Name of the signature set to use; derived from the sigcatalog}

\item{sigcatalog}{full path to signature catalog xlsx file; default is repo version}

\item{sigdbgenelist}{full path to signature DB gene list file; default is repo version}
}
\value{
the trimmed signature table
}
\description{
Read in signature catalog
}
