% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureChemicalLanePlotAcrossDatasets.R
\name{signatureChemicalLanePlotAcrossDatasets}
\alias{signatureChemicalLanePlotAcrossDatasets}
\title{Build lane plots by chemical list and signature class, across the datasets}
\usage{
signatureChemicalLanePlotAcrossDatasets(
  to.file = F,
  chemical.target = "ER",
  signature.super_class = "estrogen",
  pathset = "PathwaySet_20191107",
  method = "fc"
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}

\item{method}{Pathway scoring method in c("fc", "gsva", "mygsea")}
}
\description{
Build lane plots by chemical list and signature class, across the datasets
}
