% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureClassSummaryDotPlot.R
\name{signatureClassSummaryDotPlot}
\alias{signatureClassSummaryDotPlot}
\title{Build summary plots by signature class}
\usage{
signatureClassSummaryDotPlot(
  to.file = F,
  do.load = F,
  dataset = "DMEM_6hr_screen_normal_pe_1",
  sigset = "screen_large",
  method = "mygsea",
  sigcatalog = "signatureDB_master_catalog 2020-0505",
  hitcall.threshold = 0.5
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}

\item{do.load}{If TRUE, load hte HTTr data file}

\item{dataset}{Name of the HTTr data set}

\item{sigset}{Name of the signature set}

\item{method}{Name of the scoring method}

\item{sigcatalog}{Name of the signature catalog}

\item{hith.threshold}{Exclude rows from the dataset with hitcall less than this value}
}
\description{
Build summary plots by signature class
}
