% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureClassSummaryPlot.R
\name{signatureClassSummaryPlot}
\alias{signatureClassSummaryPlot}
\title{Build summary plots by signature class}
\usage{
signatureClassSummaryPlot(
  to.file = F,
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  sigset = "PathwaySet_20191107",
  method = "mygsea",
  hitcall.threshold = 0.5
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Build summary plots by signature class
}
