% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureConcRepFilter.R
\name{signatureConcRepFilter}
\alias{signatureConcRepFilter}
\title{Filter the conc-repons data}
\usage{
signatureConcRepFilter(
  to.file = F,
  do.plot = F,
  do.load = T,
  hccut = 0.9,
  tccut = 1.5,
  dataset = "heparg2d_toxcast_pfas_pe1_normal",
  sigset = "screen_large",
  method = "fc",
  do.pfas = F
)
}
\arguments{
\item{method}{signature scoring method in c("fc", "gsva", "mygsea")


Error bars are exp(er)*qt(.025,4) = exp(er)*2.7765
heparg2d_toxcast_pfas_pe1_normal
mcf7_ph1_pe1_normal_block_123
mcf7_ph1_pe1_normal_block_123_allPG
u2os_toxcast_pfas_pe1_normal}
}
\description{
Filter the conc-repons data
}
