% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureConcResp.R
\name{signatureConcResp}
\alias{signatureConcResp}
\title{Pathway Concentration Response (P-value)}
\usage{
signatureConcResp(
  dataset,
  sigset,
  cutoff.dataset,
  sigcatalog,
  method,
  bmr_scale = 1.349,
  mc.cores = 1,
  pval = 0.05,
  nlowconc = 2,
  aicc = F,
  minsigsize = 10,
  fitmodels = c("cnst", "hill", "gnls", "poly1", "poly2", "pow", "exp2", "exp3",
    "exp4", "exp5")
)
}
\arguments{
\item{dataset}{Name of the data set.}

\item{sigset}{Name of the signature set.}

\item{cutoff.dataset}{This is the data set name to sue when the cutoffs are taken from a different data set than
the one currently being analyzed. The reason for doing this is if the current data set is small
(small number of chemicals), and so not large enough to get a good estiamte of the underlying
noise distribution. All of the other parameters for both data sets have to be the same}

\item{method}{Pathway scoring method in c("fc", "gsva", "gsea")}

\item{bmr_scale}{bmr scaling factor. Default = 1.349}

\item{mc.cores}{Number of cores to parallelize with.}

\item{pval}{Desired cutoff p-value.}

\item{nlowconc}{Only include the lowest nlowconc concentrations for each chemical}

\item{aicc}{aicc = T uses corrected AIC to choose winning method; otherwise
regular AIC.}

\item{minsigsize}{Minimum allowed signature size. Sample/signature combinations
with less than this number of non-missing l2fc's will be discarded.}

\item{fitmodels}{Vector of model names to use. Probably should include "cnst".}
}
\value{
If to.file = T, nothing. If to.file = F, dataframe with signature CR
  output.
}
\description{
Performs signature concentration response using p-value based cutoffs.
}
\details{
dataset should have already been scored using signatureScore
and the given sigset and method. This function prepares signatureScore output
for CR processing, calls signatureConcRespCore_pval, formats the output,
saves it to disk
}
