% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureConcRespFilter.R
\name{signatureConcRespFilter}
\alias{signatureConcRespFilter}
\title{Filter the conc-response data for jsut the most potent results
and plot the conc-response curves if desired}
\usage{
signatureConcRespFilter(
  to.file = F,
  do.plot = F,
  do.load = T,
  hccut = 0.9,
  tccut = 1.5,
  dataset = "heparg2d_toxcast_pfas_pe1_normal",
  sigset = "screen_large",
  method = "fc",
  do.pfas = F
)
}
\arguments{
\item{to.file}{If TRUE, send plots to a file}

\item{do.plot}{If TRUE do the plotting}

\item{hccut}{Exclude rows with hitcall below this value}

\item{tccut}{Exclude rows with top_over_cutoff below this value}

\item{dataset}{Dataset to use}

\item{sigset}{Signature set to use}

\item{method}{signature scoring method in c("fc", "gsva", "mygsea")}

\item{If}{TRUE, load the data file}

\item{do.pfas=F}{Error bars are exp(er)*qt(.025,4) = exp(er)*2.7765}
}
\description{
Filter the conc-response data for jsut the most potent results
and plot the conc-response curves if desired
}
