% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureConcRespNullDistThreshold.R
\name{signatureConcRespNullDistThreshold}
\alias{signatureConcRespNullDistThreshold}
\title{Export the Null set thresholds}
\usage{
signatureConcRespNullDistThreshold(
  sigset = "screen_large",
  dataset = "mcf7_ph1_pe1_normal_all_pg",
  method = "fc",
  pval = 0.05
)
}
\arguments{
\item{sigset}{Name of the signature set.}

\item{dataset}{Name of the data set.}

\item{method}{Pathway scoring method in c("fc", "gsva", "gsea")}

\item{pval}{Desired cutoff p-value.}
}
\description{
Null dataset and dataset should have already been scored using signatureScore
and the given sigset and method.
}
