% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureConcRespPlotWrapper.R
\name{signatureConcRespPlotWrapper}
\alias{signatureConcRespPlotWrapper}
\title{Wrapper for all of the conc-response plotting}
\usage{
signatureConcRespPlotWrapper(
  sigset,
  dataset,
  sigcatalog,
  method,
  bmr_scale = 1.349,
  mc.cores = 20,
  do.load = T,
  pval = 0.05,
  plotrange = c(1e-04, 100)
)
}
\arguments{
\item{sigset}{Name of the signature set.}

\item{dataset}{Name of the data set.}

\item{sigcatalog}{Name of the signature catalog}

\item{method}{Pathway scoring method in c("fc", "gsva", "gsea")}

\item{bmr_scale}{bmr scaling factor. Default = 1.349}

\item{mc.cores}{Number of cores to parallelize with.}

\item{do.load}{If TRUE, load the SIGNATURE_CR file, otherwiseassume that it is in memory
to.file to.file = T saves the output to a file; otherwise it's returned.}

\item{pval}{Desired cutoff p-value.}

\item{plotrange}{The x-range of the plot as a vector of 2 elements, this can be changed for special cases, but defaults to 0.001 to 100}
}
\description{
Wrapper for all of the conc-response plotting
}
