% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureConcRespToZ.R
\name{signatureConcRespToZ}
\alias{signatureConcRespToZ}
\title{Convert the conc-response data to a z score}
\usage{
signatureConcRespToZ(
  do.load = T,
  mc.cores = 2,
  dataset = "heparg2d_toxcast_pfas_pe1_normal",
  sigset = "screen_large",
  method = "fc",
  celltype = "HepaRG",
  hccut = 0.95,
  tccut = 1.5
)
}
\arguments{
\item{do.load}{If TRUE, load hte large HTTr data set}

\item{mc.cores}{NUmber of cores to use in multi-core mode=2,}

\item{dataset}{Name of the HTTr data set being used}

\item{sigset}{Name of the signature set used}

\item{method}{Scoring method used}

\item{celltype}{name of cell type ebing used}

\item{hccut}{Exclude signature rows with hitcall less than this value}

\item{tccut}{Exclude signature rows with top_over_cutoff less than this value}
}
\description{
Convert the conc-response data to a z score
}
