% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureDirectionPlot.R
\name{signatureDirectionPlot}
\alias{signatureDirectionPlot}
\title{Plot the cummulative distribution functions of the up and down direction signatures}
\usage{
signatureDirectionPlot(
  to.file = T,
  do.load = F,
  dataset = "MCF7_pilot_PRF_6hr_pilot_normal_pe_1",
  sigset = "screen_large",
  method = "gsea",
  celltype = "MCF7",
  hccut = 0.9,
  tccut = 1
)
}
\arguments{
\item{to.file}{If TRUE, send the plots to a file}

\item{do.load}{If TRUE, load hte large HTTr data set into memory}

\item{dataset}{Name of the HTTr data set}

\item{sigset}{Name of the signature set}

\item{method}{Scoring method}

\item{celltype}{Name of the cell type}

\item{hccut}{Exclude rows in the data set with hitcall less than this value}

\item{tccut}{Exclude rows in the data set with top_over_cutoff less than this value}

\item{sigcatalog}{Name of the signature catalog to use}

\item{cutoff}{The minimum number of signatures hat have to be active in a super
target for the super target to be considered active. Default is 5}

\item{minconc}{Minimum concentration used in the plots}

\item{maxconc}{Maximum concentration used in the plots

After running this, run the following ...
superTargetPODplot
superTargetStats}
}
\description{
Plot the cummulative distribution functions of the up and down direction signatures
}
