% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureGeneHM.R
\name{signatureGeneHM}
\alias{signatureGeneHM}
\title{Build lane plots by chemical list and signature class, across the datasets}
\usage{
signatureGeneHM(
  to.file = F,
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  chemical.target = "ER",
  signature.super_target = "estrogen",
  sigset = "pilot_tiny",
  sigcatalog = "signatureDB_master_catalog 2020-01-31",
  method = "mygsea",
  threshold = 0.5
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Build lane plots by chemical list and signature class, across the datasets
}
