% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureHeatmap.R
\name{signatureHeatmap}
\alias{signatureHeatmap}
\title{Create heatmaps of the htpp data frames}
\usage{
signatureHeatmap(
  to.file = F,
  pathset = "PathwaySet_20191025",
  dataset = "DMEM_6hr_pilot_normal_00",
  method = "fc",
  conthits = T,
  nametag = NULL,
  pval = 0.05
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}

\item{dataset}{The set of data to be included}
}
\description{
Create heatmaps of the htpp data frames
}
