% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signaturePOD.BMRcompare.R
\name{signaturePOD.BMRcompare}
\alias{signaturePOD.BMRcompare}
\title{Compare the PODs with different BMR values}
\usage{
signaturePOD.BMRcompare(
  to.file = F,
  dataset = "mcf7_ph1_pe1_normal_block_123",
  sigset = "screen_large",
  method = "fc",
  bmr_scale = 1,
  hccut = 0.9
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}

\item{dataset}{Name of data set.}

\item{sigset}{Name of signature set.}

\item{method}{Pathway scoring method}

\item{bmr_scale}{bmr scaling factor. Default = 1.349}

\item{hccut}{Remove rows with hitcall less than this value}
}
\description{
Compare the PODs with different BMR values
}
