% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signaturePOD.R
\name{signaturePOD}
\alias{signaturePOD}
\title{Calculate PODs at the signature level}
\usage{
signaturePOD(
  do.load = F,
  sigset = "screen_large",
  dataset = "MCF7_pilot_DMEM_6hr_pilot_normal_pe_1",
  method = "gsea",
  hccut = 0.9,
  cutoff = 3,
  condition = "all"
)
}
\arguments{
\item{sigset}{Name of signature set.}

\item{dataset}{Name of data set.}

\item{method}{Pathway scoring method in c("fc", "gsva", "gsea")}

\item{hccut}{Remove rows with hitcall less than this value}

\item{do.laod}{If TRUE, load the input data into memory}

\item{bmr_scale}{bmr scaling factor. Default = 1.349}
}
\description{
Calculate PODs at the signature level
}
