% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signaturePODsummary.R
\name{signaturePODsummary}
\alias{signaturePODsummary}
\title{Summarize the POD overlap with ToxCast}
\usage{
signaturePODsummary(
  sigset = "pilot_large_all_100CMAP",
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  method = "mygsea"
)
}
\arguments{
\item{sigset}{THe name of the signature set}

\item{dataset}{Name of the HTTr data set}

\item{method}{THe signature scoring method}
}
\description{
Summarize the POD overlap with ToxCast
}
