% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureRank.DESEQ2.R
\name{signatureRank.DESEQ2}
\alias{signatureRank.DESEQ2}
\title{Get the signature ranks for chemicals}
\usage{
signatureRank.DESEQ2(
  to.file = F,
  dataset1 = "DMEM_6hr_pilot_none_pe_1",
  dataset2 = "DMEM_6hr_pilot_normal_pe_1",
  pathset = "PathwaySet_20191107",
  method = "gsva",
  cutoff = 0.5
)
}
\description{
Get the signature ranks for chemicals
}
