% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureScoreCoreFC.R
\name{signatureScoreCoreFC}
\alias{signatureScoreCoreFC}
\title{Signature Score Core - FC}
\usage{
signatureScoreCoreFC(
  fcmat,
  sigset,
  chem_dict,
  signature_data,
  ngenemax = NULL,
  verbose = F
)
}
\arguments{
\item{fcmat}{Expects FCMAT2. Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{sigset}{Name of signature set.}

\item{chem_dict}{Expects CHEM_DICT object. Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}

\item{signature_data}{Named list of gene name vectors. Each element is one
signature, defined by the genes it contains.}

\item{ngenemax}{If ngene is not NULL, then only the most extreme n genes of the
signature will be used for the "in" set}

\item{verbose}{If TRUE, write extra diagnostic output}
}
\value{
Dataframe with one row per chemical/conc/signature combination. Columns
  are: sample_id, dtxsid, casrn, name, time, conc, sigset,
  signature, size (signature size accounting for missing values), mean_fc_scaled_in,
  mean_fc_scaled_out, signature_score.
}
\description{
Computes fold change signature scores.
}
\details{
This fast implementation of fold change signature scores uses matrix
multiplication. The score is simply: mean(fold change of genes in signature) -
mean(fold change of genes outside signature).
}
