% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureScoreCoreGSVA.R
\name{signatureScoreCoreGSVA}
\alias{signatureScoreCoreGSVA}
\title{Signature Score Core - GSVA}
\usage{
signatureScoreCoreGSVA(
  sk.list,
  sigset = "FILTERED",
  fcmat,
  chem_dict,
  signature_data,
  mc.cores = 1
)
}
\arguments{
\item{sk.list}{Sample keys to use; should correspond to fcmat rownames.}

\item{sigset}{Name of signature set.}

\item{fcmat}{Expects FCMAT2. Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{chem_dict}{Expects CHEM_DICT object. Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}

\item{signature_data}{Named list of gene name vectors. Each element is one
signature, defined by the genes it contains.}

\item{mc.cores}{Number of cores to use. Parallelization is performed
by gsva itself.}
}
\value{
Dataframe with one row per chemical/conc/signature combination. Columns
  are: sample_id, dtxsid, casrn, name, time, conc, sigset, signature, size, signature_score
}
\description{
Computes GSVA signature scores.
}
\details{
This function is a wrapper for GSVA with Gaussian cumulative distribution function (cdf) kernels.
}
