% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureScoreMerge.R
\name{signatureScoreMerge}
\alias{signatureScoreMerge}
\title{Merge the up and down halves of the pathway data}
\usage{
signatureScoreMerge(sigset, sigcatalog, dataset, method)
}
\arguments{
\item{sigset}{Name of the signature set.}

\item{dataset}{Name of the data set.}

\item{method}{Pathway scoring method in c("fc", "gsva", "gsea")}

\item{sigcatlog}{Name of the catalog file}
}
\value{
nothing
}
\description{
Merge the up and down halves of the pathway data
}
