% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureScoreVariance.R
\name{signatureScoreVariance}
\alias{signatureScoreVariance}
\title{Visualize the variance by signature - this supplants signatureVariance.R}
\usage{
signatureScoreVariance(
  to.file = F,
  do.load = F,
  sigset = "pilot_large_all_CMAP",
  sigcatalog = "signatureDB_master_catalog 2020-01-31",
  basedir = "../input/fcdata/",
  dataset = "DMEM_6hr_screen_normal_pe_1",
  method = "mygsea"
)
}
\arguments{
\item{min.ngene}{Signatures will only be saved if the number of genes is >= this value}

\item{max.ngene}{Signatures will only be saved if the number of genes is <= this value}
}
\value{
No output.
}
\description{
Visualize the variance by signature - this supplants signatureVariance.R
}
