% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSlice.R
\name{signatureSlice}
\alias{signatureSlice}
\title{Look at concentration-slides of gene CR data to understand where burst starts}
\usage{
signatureSlice(
  to.file = F,
  do.load = F,
  dataset = "mcf7_ph1_pe1_normal_block_123_allPG",
  sigcatalog = "signatureDB_master_catalog 2021-09-29",
  sigset = "screen_large",
  method = "gsea",
  celltype = "MCF7",
  hccut = 0.9,
  minhit = 10,
  tccut = 0.9
)
}
\arguments{
\item{to.file}{If TRUE, send the plots to a file}

\item{do.load}{If TRUE, load hte large HTTr data set into memory}

\item{dataset}{Name of the HTTr data set}

\item{sigcatalog}{Name of the signature catalog to use}

\item{sigset}{Name of the signature set}

\item{method}{Scoring method}

\item{celltype}{Name of the cell type}

\item{hccut}{Exclude rows in the data set with hitcall less than this value}

\item{tccut}{Exclude rows in the data set with top_over_cutoff less than this value}

\item{cutoff}{The minimum number of signatures hat have to be active in a super
target for the super target to be considered active. Default is 5}

\item{minconc}{Minimum concentration used in the plots}

\item{maxconc}{Maximum concentration used in the plots

After running this, run the following ...
superTargetPODplot
superTargetStats}
}
\description{
Look at concentration-slides of gene CR data to understand where burst starts
}
