% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureTestMerging.R
\name{signatureTestMerging}
\alias{signatureTestMerging}
\title{Examine the merging code effect}
\usage{
signatureTestMerging(
  to.file = F,
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  sigset = "pilot_tiny",
  method = "mygsea",
  cname = "Fulvestrant"
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Examine the merging code effect
}
