% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureToBMDExpressFormat.R
\name{signatureToBMDExpressFormat}
\alias{signatureToBMDExpressFormat}
\title{Map the signatures ot the BMDExpress format}
\usage{
signatureToBMDExpressFormat(
  sigset = "pilot_tiny",
  sigcatalog = "signatureDB_master_catalog 2020-01-31"
)
}
\arguments{
\item{sigset}{Name of the signature set.}

\item{sigcatlog}{Nmae of the catalog file}
}
\value{
the trimmed signature table
}
\description{
Map the signatures ot the BMDExpress format
}
