% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAccumNullPlot.R
\name{smoothecdf}
\alias{smoothecdf}
\title{Smooth ECDF}
\usage{
smoothecdf(x, mymat, verbose = F, bmdrange = c(0.001, 100))
}
\arguments{
\item{x}{ECDF plotting location x-values.}

\item{mymat}{Dataframe containing bmd10, bmdu, bmdl, and hitcall columns.}

\item{verbose}{verbose = F suppresses both bounds NA warning.}

\item{bmdrange}{Maximum expected BMD range. The farthest value from the bmd10
is used to compute standard deviation of gaussian when both bounds are missing.}
}
\value{
Outputs a vector corresponding to the locations in x.
}
\description{
Converts a data frame containing bmd10, bmdl, bmu, to a smooth ecdf.
}
\details{
Models each bmd as a gaussian with mean bmd10 uses bmdl (bmdu if bmdl is na) 
to compute sd. Each gaussian is scaled by the hitcall.
}
\examples{
x = 10^(-50:50/30)
mymat =data.frame(list(bmd10 = c(.1,1,10), bmdl = c(.05,NA,NA), 
  bmdu = c(.6,1.5,NA), hitcall = c(1,1,1)))
out = smoothecdf(x, mymat)
plot(log10(x),out, type = "l")
mymat$hitcall = c(1,.5,0)
out2 = smoothecdf(x, mymat)
plot(log10(x),out2, type = "l")
}
