% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stressSignatureHM.R
\name{stressPathwayHM}
\alias{stressPathwayHM}
\title{Build a heatmap of the stress genes}
\usage{
stressPathwayHM(
  to.file = F,
  dataset = "DMEM_6hr_pilot_normal_pe_0",
  pathset = "PathwaySet_20191107",
  threshold = 0.5,
  method = "fc"
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Build a heatmap of the stress genes
}
