% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superTargetBoxplot.R
\name{superTargetBoxplot}
\alias{superTargetBoxplot}
\title{Generate chemical-wise boxplot of the BMD distributions by super_target}
\usage{
superTargetBoxplot(
  to.file = T,
  do.load = T,
  dataset = "u2os_toxcast_pfas_pe1_normal_v2_refchems",
  sigcatalog = "signatureDB_master_catalog 2021-09-29",
  sigset = "screen_large",
  method = "gsea",
  celltype = "U2OS",
  hccut = 0.9,
  tccut = 1,
  cutoff = 3,
  minconc = 0.001,
  maxconc = 100,
  chemfile = NULL
)
}
\arguments{
\item{to.file}{If TRUE, send the plots to a file}

\item{do.load}{If TRUE, load hte large HTTr data set into memory}

\item{dataset}{Name of the HTTr data set}

\item{sigcatalog}{Name of the signature catalog to use}

\item{sigset}{Name of the signature set}

\item{method}{Scoring method}

\item{celltype}{Name of the cell type}

\item{hccut}{Exclude rows in the data set with hitcall less than this value}

\item{tccut}{Exclude rows in the data set with top_over_cutoff less than this value}

\item{cutoff}{The minimum number of signatures hat have to be active in a super
target for the super target to be considered active. Default is 5}

\item{minconc}{Minimum concentration used in the plots}

\item{maxconc}{Maximum concentration used in the plots

After running this, run the following ...
superTargetPODplot
superTargetStats}
}
\description{
Generate chemical-wise boxplot of the BMD distributions by super_target
}
