% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superTargetStats.R
\name{superTargetStats}
\alias{superTargetStats}
\title{Generate hit statistics by super_target}
\usage{
superTargetStats(
  do.load = F,
  dataset = "heparg2d_toxcast_pfas_pe1_normal_refchems",
  sigset = "pilot_tiny",
  method = "fc",
  celltype = "HepaRG",
  hccut = 0.95,
  tccut = 1.5,
  cutoff = 5
)
}
\arguments{
\item{do.load}{If TRUE, Load the large input data file}

\item{dataset}{Name of the HTTr data set}

\item{sigset}{Name of the signature set}

\item{method}{Scoring method}

\item{celltype}{Name of the cell type}

\item{hccut}{Exclude rows in the data set with hitcall less than this value}

\item{tccut}{Exclude rows in the data set with top_over_cutoff less than this value}

\item{cutoff}{The minimum number of signatures hat have to be active in a super
target for the super target to be considered active. Default is 5}
}
\description{
Generate hit statistics by super_target
}
