% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxT.R
\name{TxT}
\alias{TxT}
\title{Calculate several statistics on a 2 x 2 matrix}
\usage{
TxT(tp, fp, fn, tn, do.p = TRUE, rowname = NA)
}
\arguments{
\item{tp}{number of true positives}

\item{fp}{number of false positives}

\item{fn}{number of false negatives}

\item{tn}{number of true negatives}

\item{do.p}{if TRUE, calcualte an exact p-value}

\item{rowname}{if not NA, adda column to the output with this rowname

Returns:
  a list of the results
  a: TP
  b: FP
  c: FN
  d: TN
  sens: sensitivity
  spec: specificity
  ba: Balanced Accuracy
  accuracy: Accuracy
  relative.risk: Relative Risk
  odds.ratio: Odds Ratio
  or.ci.lwr: lower confidence interval of the Odds Ratio
  or.ci.upr: upper confidence interval of the Odds Ratio
  ppv: Positive Predictive Value
  npv: Negative Predictive Value
  p.value: Chi-squared p-value
  F1: 2TP/(2TP+FP+FN)

  sval: All of the results as a tab-delimited string
  title: the title of the results as a tab-delimited string
 mat: The results as a 1-row data frame
 @export}
}
\description{
Calculate several statistics on a 2 x 2 matrix
}
