% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFCMAT1.fromDB.R
\name{buildFCMAT1.fromDB}
\alias{buildFCMAT1.fromDB}
\title{Build the FCMAT1 data set}
\usage{
buildFCMAT1.fromDB(
  dataset = "mcf7_ph1_pe1_normal_block_123_excludePG",
  dir = "../input/fcdata/new_versions/",
 
    infile = "httr_mcf7_ph1_bl123_FCmat1_meanncnt0_5-plateteffect_1-shrinkage_normal.RData",
  pg.filter.file = "httr_mcf7_ph1_flagged_pg_block_123_exclude.xlsx",
  do.load = T
)
}
\arguments{
\item{dataset}{The name to give to the data set}

\item{dir}{The directory from which to read all of the raw files}

\item{infile}{The nae of the input file}

\item{pg.filter.file}{An optional file to use in filtering out bad plate groups}

\item{do.load}{If TRUE, read the large input data file into memory}
}
\value{
A file with the FCMAT1 data is written to "../input/fcdata/FCMAT1_",dataset,".RData"
}
\description{
version to start with Logan's database export
The difference between this version and the original is that there are extra columns
The function just changes one column name and writes the file to a standard name and place
}
