% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFCMAT2.fromDB.R
\name{buildFCMAT2.fromDB}
\alias{buildFCMAT2.fromDB}
\title{Transpose and filter the fold change matrix FCMAT1 in long format into a
gene x sample format.}
\usage{
buildFCMAT2.fromDB(
  dataset = "tox21_cpp5_u2os_pe1_normal",
  time = 24,
  media = "DMEM",
  dir = "../input/fcdata/",
  method = "gene",
  do.read = T
)
}
\arguments{
\item{dataset}{The name to give to the data set}

\item{time}{The time in hours that the chemical dosing was run}

\item{media}{THe name of the media used}

\item{dir}{The directory from which to read all of the raw files}

\item{method}{Either "gene" or "probe"}

\item{do.read}{If TRUE, read in the FCMAT1 file and place in a global.}
}
\value{
Global variables are created for the FC matrix (FCMAT2), the SE matrix (SEMAT2)
and the chemical dictionary (CHEM_DICT) which translates form the sample key
(sample_id_conc_time) to the individual components
}
\description{
Transpose and filter the fold change matrix FCMAT1 in long format into a
gene x sample format.
}
