% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFCMAT2.R
\name{buildFCMAT2}
\alias{buildFCMAT2}
\title{Transpose and filter the fold change matrix FCMAT1 in long format into a
gene x sample format.
Function returns FCMAT2 and CHEM_DICT as a list R object}
\usage{
buildFCMAT2(
  FCMAT1 = NULL,
  time = 24,
  media = "DMEM",
  output_dir = "../input/fcdata/",
  method = "gene",
  sample_type = "test sample",
  writing_to_disk = FALSE
)
}
\arguments{
\item{FCMAT1}{The input data structure that contains the data frame from the httrlib buildFCMAT1() function or provided by the user in the same format.
FCMAT1 should be a long format table with rows denoting genes/probes and an 'l2fc' column which contains log2 fold-change data from DESeq2 or another method.}

\item{time}{The time in hours that the chemical dosing was run}

\item{media}{The name of the media used}

\item{output_dir}{The directory from which to read all of the raw files}

\item{method}{Either "gene" or "probe_id". Specifying "gene" will average probes that target the same gene, whereas "probe_id" will keep all probes separate when making the FCMAT2 table.}

\item{sample_type, }{either "test sample" or "reference chemical" will determine what sample type the FCMAT2/CHEM_DICT tables will include}

\item{writing_to_disk}{dictates whether all results should be saved to disk (path is defined by output_dir parameter)}
}
\value{
Global variables are created for the FC matrix (FCMAT2)
and the chemical dictionary (CHEM_DICT) which translates from the sample key
(sample_id_conc_time) to the individual components

list consisting of matrices FCMAT2 and CHEM_DICT
}
\description{
Transpose and filter the fold change matrix FCMAT1 in long format into a
gene x sample format.
Function returns FCMAT2 and CHEM_DICT as a list R object
}
