% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStudyChemicalMap.R
\name{buildStudyChemicalMap}
\alias{buildStudyChemicalMap}
\title{Build a catalog of the chemicals in a dataset}
\usage{
buildStudyChemicalMap(dataset = "DMEM_6hr_screen_normal_pe_1")
}
\arguments{
\item{dataset}{The name of the HTTr dataset}
}
\value{
No output.
}
\description{
Build a catalog of the chemicals in a dataset
}
