% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffCalc.R
\name{cutoffCalc}
\alias{cutoffCalc}
\title{Calculate the signature-wise cutoffs based on the Setzer method
which does not break any correlations between genes}
\usage{
cutoffCalc(
  basedir = "../input/fcdata/",
  dataset = "mcf7_ph1_pe1_normal_block_123_allPG",
  sigcatalog = "signatureDB_master_catalog 2021-05-10",
  sigset = "screen_large",
  method = "fc",
  p = 0.05,
  seed = 12345,
  nlowconc = 2,
  dtxsid.exclude = NULL,
  do.load = T,
  do.cov = T,
  do.compare = F,
  to.file = F,
  verbose = F
)
}
\arguments{
\item{basedir}{Directory that holds FCMAT2 and CHEM_DICT files.}

\item{dataset}{Name of actual dataset to base null data on.}

\item{sigcatalog}{The name of the signature catalog to use}

\item{sigset}{THe signature set}

\item{method}{The scoring method, either fc or gsea}

\item{p}{The p-value for the null distribution}

\item{seed}{Random seed.}

\item{nlowconc}{If not NULL, only include the lowest nlowconc concentrations for each chemical}

\item{dtxsid.exclude}{dtxsids to exclude, default NULL}

\item{do.load}{If TRUE, reload the FCMAT2 matrix, signature catalog and chemical dictionary, and store in globals}

\item{do.cov}{If TRUE, calculate the covariance matrix and store in a global}

\item{do.compare}{If TRUE, compare the cutoffs with those from the original method with no gene-gene correlation}

\item{to.file}{If TRUE, and do.compare=TRUE, send a plot of the comparison to a file}

\item{verbose}{If TRUE, write a line for each signature to show progress.}
}
\value{
No output.
}
\description{
Calculate the signature-wise cutoffs based on the Setzer method
which does not break any correlations between genes
}
