% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.driver.R
\name{driver.driver}
\alias{driver.driver}
\title{Code to run all calculations}
\usage{
driver.driver(
  pathset = "PathwaySet_20200108",
  method = "mygsea",
  nrandom.chems = 1000,
  mc.cores = 30,
  step1 = F,
  step2 = F,
  step3 = F,
  do.log = F
)
}
\arguments{
\item{method}{Pathway scoring method in c("fc", "gsva", "mygsea")}
}
\description{
Code to run all calculations
}
