% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{Code to run all calculations}
\usage{
driver(
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  sigcatalog = "signatureDB_master_catalog 2020-01-31",
  sigset = "pilot_tiny",
  nrandom.chems = 1000,
  mc.cores = 1,
  method = "mygsea",
  do.build.fcmat1.all = F,
  do.build.fcmat2.all = F,
  do.build.random = F,
  do.run.random = F,
  do.run.all = F,
  do.signature.summary.plot = F,
  do.signature.pod = F,
  do.signature.pod.laneplot = F,
  do.all = F
)
}
\arguments{
\item{method}{signature scoring method in c("fc", "gsva", "mygsea")}
}
\description{
Code to run all calculations
}
