% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{Code to run all signature concentration-response calculations}
\usage{
driver(
  dataset = "heparg2d_toxcast_pfas_pe1_normal",
  sigcatalog = "signatureDB_master_catalog 2021-05-10",
  sigset = "screen_large",
  cutoff.dataset = NULL,
  normfactor = 7500,
  mc.cores = 25,
  bmr_scale = 1.349,
  pval = 0.05,
  nlowconc = 2,
  hccut = 0.95,
  tccut = 1.5,
  plotrange = c(1e-04, 100),
  method = "fc",
  celltype = "HepaRG",
  do.conc.resp = T,
  do.scr.plots = T,
  do.signature.pod = T,
  do.supertarget.boxplot = T,
  do.all = F
)
}
\arguments{
\item{dataset}{Name of the data set, produced by buildFCMAT2}

\item{sigcatalog}{Name of the signature catalog}

\item{sigset}{Name if the signature set. THis corresponds to a column in the signature catalog file}

\item{cutoff.dataset}{This is the data set name to sue when the cutoffs are taken from a different data set than
the one currently being analyzed. The reason for doing this is if the current data set is small
(small number of chemicals), and so not large enough to get a good estiamte of the underlying
noise distribution. All of the other parameters for both data sets have to be the same}

\item{normfactor}{Normalization factor for the conc-reap plots, default is 7500}

\item{mc.cores}{Number of cores for parallel processing. Only works under Linux}

\item{bmr_scale}{Scaling factor from the NULL SD to BMD, default is 1.349}

\item{pval}{Threshold for cutoff distribution confidence interval. Default=0.05 indicates a 95% CI for the baseline distribution}

\item{nlowconc}{Only include the lowest nlowconc concentrations for each chemical}

\item{hccut}{The threshold for signatures to be called a hit, default=0.95,}

\item{tccut}{The threshold for top/cutoff o be a hit, default =1.5}

\item{plotrange}{The concentration range for the conc-resp plots in uM, default is c(0.0001,100),}

\item{method}{signature scoring method in c("fc", "gsva", "gsea"), default is fc}

\item{celltype}{Name of the cull type, e.g. MCF7}

\item{do.conc.resp}{If true, run the concentration-response calculations}

\item{do.scr.plots}{If TRUE, generate the signature concentration response plots}

\item{do.signature.pod}{If TRUE, generate the signature PODs}

\item{do.supertarget.boxplot}{If TRUE, generate the super target box plots}

\item{do.all}{If TRUE, do all steps from do.build.random to the end


Available data sets
* heparg2d_toxcast_pfas_pe1_normal
* mcf7_ph1_pe1_normal_block_123_allPG
* mcf7_ph1_pe1_normal_block_123_excludePG
* u2os_toxcast_pfas_pe1_normal
* PFAS_HepaRG
* PFAS_U2OS
* u2os_pilot_pe1_normal_null_pilot_lowconc
* u2os_toxcast_pfas_pe1_normal_refchems
* heparg2d_toxcast_pfas_pe1_normal_refchems
* DMEM_6hr_pilot_normal_pe_1 - MCF7 pilot}

\item{do.signature.summary.plot}{if TRUE, generate the summary plots}

\item{do.signature.pod.laneplot}{If TRUE, generate the signature lane plots (only useful for small sets of chemicals)}
}
\description{
Code to run all signature concentration-response calculations
}
