% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSignatureCutoffs.R
\name{exportSignatureCutoffs}
\alias{exportSignatureCutoffs}
\title{Export the signature-wise cutoffs}
\usage{
exportSignatureCutoffs(signature_cr)
}
\arguments{
\item{signature_cr}{dataframe returned by the signatureConcResp function}
}
\description{
Export the signature-wise cutoffs
}
