% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneBaseMeanDist.R
\name{geneBaseMeanDist}
\alias{geneBaseMeanDist}
\title{get the bse mean distributio nfor each gene}
\usage{
geneBaseMeanDist(
  to.file = F,
  do.read = F,
  dataset = "DMEM_6hr_screen_normal_pe_1"
)
}
\arguments{
\item{min.ngene}{Signatures will only be saved if the number of genes is >= this value}

\item{max.ngene}{Signatures will only be saved if the number of genes is <= this value}
}
\value{
No output.
}
\description{
get the bse mean distributio nfor each gene
}
