% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneConcRespPlotWrapper.R
\name{geneConcRespPlotWrapper}
\alias{geneConcRespPlotWrapper}
\title{Wrapper for all of the conc-response plotting o genes}
\usage{
geneConcRespPlotWrapper(
  dataset = "tox21_cpp5_heparg_pe1_normal",
  mc.cores = 20,
  do.load = T,
  to.file = F,
  pval = 0.05,
  plotrange = c(1e-04, 100),
  onefile = T,
  chemfile = NULL
)
}
\arguments{
\item{dataset}{Name of the data set.}

\item{mc.cores}{Number of cores to parallelize with.}

\item{do.load}{If TRUE, load the SIGNATURE_CR file, otherwiseassume that it is in memory}

\item{to.file}{to.file = T saves the output to a file; otherwise it's returned.}

\item{pval}{Desired cutoff p-value.}

\item{plotrange}{The x-range of the plot as a vector of 2 elements, this can be changed for special cases, but defaults to 0.001 to 100}

\item{onefile}{If TRUE, put all plots into one file, instead of one filer per chemical}

\item{chemfile}{A file of chemicals to use. If NULL, plot all chemicals}
}
\description{
Wrapper for all of the conc-response plotting o genes
}
