% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayConcResp_pval.R
\name{plotouter}
\alias{plotouter}
\title{Plot Outer}
\usage{
plotouter(proper_name, PATHWAY_CR, foldname)
}
\arguments{
\item{proper_name}{Chemical name to be used in file name.}

\item{PATHWAY_CR}{Dataframe output of pathwayConcResp_pval.}

\item{foldname}{Folder name for output file.}
}
\value{
No output.
}
\description{
Calls pathwayConcResp plotting function.
}
\details{
Calls pathwayConcResp plotting function for one chemical and every pathway.
Saves a single pdf to disk for the given chemical containing every pathway 
CR plot.
}
