% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomdata.R
\name{randomdata}
\alias{randomdata}
\title{Randomized Null Data}
\usage{
randomdata(
  basedir = "../input/fcdata/",
  dataset = "u2os_pilot_pe1_normal_null_pilot_lowconc",
  nchem = 1000,
  seed = 12345,
  maxconc = 1e+06,
  nlowconc = 2,
  dtxsid.exclude = NULL
)
}
\arguments{
\item{basedir}{Directory that holds FCMAT2 and CHEM_DICT files.}

\item{dataset}{Name of actual dataset to base null data on.}

\item{nchem}{Number of null chemicals. Number of null samples is approximately
eight times this value.}

\item{seed}{Random seed.}

\item{maxconc}{Only use concentrations less than maxconc, default 1000000}

\item{nlowconc}{If not NULL, only include the loest nlowconc concentrations for eahc chemical}

\item{dtxsid.exclude}{dtxsids to exclude, default NULL
for U2OS pilot dtxsid.exclude=c('DTXSID9020031','DTXSID0040464','DTXSID5023582')}
}
\value{
No output.
}
\description{
Generate randomized null data based on actual data.
}
\details{
New FCMAT2 and CHEM_DICT files corresponding to the null dataset are written
to disk in the basedir folder. The nullset name is paste0(dataset, "_", nchem).
Randomization is performed by sampling the quantile function for each gene in
the actual data. The nullset will have roughly the same distribution of values
for each gene in the actual data,
}
