% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAllSignatureCR.R
\name{runAllSignatureCR}
\alias{runAllSignatureCR}
\title{Run All Pathway Concentration Response (P-Value)}
\usage{
runAllSignatureCR(
  dataset,
  sigset,
  sigcatalog = "../inst/extdata/signatureDB_master_catalog_2022-05-16.xlsx",
  method,
  bmr_scale = 1.349,
  normfactor = 7500,
  minsigsize = 10,
  pval = 0.05,
  nlowconc = 2,
  mc.cores = 1,
  fitmodels = c("cnst", "hill", "poly1", "poly2", "pow", "exp2", "exp3", "exp4", "exp5"),
  FCMAT2,
  CHEM_DICT,
  sigdbgenelist = "../inst/extdata/signatureDB_genelists.RDS"
)
}
\arguments{
\item{dataset}{Name of data set.}

\item{sigset}{Name of signature set.}

\item{sigcatalog}{Name of the signature catalog}

\item{method}{Pathway scoring method in c("fc", "gsva", "gsea")}

\item{bmr_scale}{benchmark response (bmr) scaling factor. Default = 1.349}

\item{normfactor}{Factor to scale the native units up by to get onto a reasonable plotting value (~ -1 to 1)}

\item{minsigsize}{Minimum signature size.}

\item{pval}{P-value to use for noise estimation.}

\item{nlowconc}{Only include the lowest nlowconc concentrations for each chemical}

\item{mc.cores}{Vector with two values: number of cores to use for signature
scoring and number of cores to use for CR. CR can usually handle the maximum
number, but gsva scoring might require a smaller number to avoid memory
overflow.}

\item{fitmodels}{Vector of model names to run conc/resp with. "cnst" should
always be chosen.}

\item{FCMAT2}{Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{CHEM_DICT}{Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}

\item{sigdbgenelist}{full path to signature DB gene list file; default is repo version}
}
\value{
dataframe with signature concentration-response results
}
\description{
Driver for signature scoring and concentration response (CR).
}
