% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAllSignatureCR.R
\name{runAllSignatureCR}
\alias{runAllSignatureCR}
\title{Run All Pathway Concentration Response (P-Value)}
\usage{
runAllSignatureCR(
  basedir = "../input/fcdata/",
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  sigset,
  sigcatalog,
  method = "mygsea",
  minsigsize = 10,
  conthits = T,
  nullset,
  do.plot = T,
  pval = 0.05,
  mc.cores = c(1, 1),
  fitmodels = c("cnst", "hill", "poly1", "poly2", "pow", "exp2", "exp3", "exp4", "exp5")
)
}
\arguments{
\item{basedir}{Folder that stores FCMAT2 and CHEM_DICT files.}

\item{dataset}{Name of data set.}

\item{sigset}{Name of signature set.}

\item{method}{Pathway scoring method in c("fc", "gsva", "mygsea")}

\item{minsigsize}{Minimum signature size.}

\item{conthits}{conthits = T uses continous hitcall; conthits = F uses discrete
hitcalls.}

\item{nullset}{Name of null dataset. Set nullset = NULL to skip CR.}

\item{do.plot}{do.plot = T generates a CR plot for every sample/signature
combination.}

\item{pval}{P-value to use for noise estimation.}

\item{mc.cores}{Vector with two values: number of cores to use for signature
scoring and number of cores to use for CR. CR can usually handle the maximum
number, but gsva scoring might require a smaller number to avoid memory
overflow.}

\item{fitmodels}{Vector of model names to run conc/resp with. "cnst" should
always be chosen.}
}
\value{
No output.

remove gnls from default set
}
\description{
Driver for signature scoring and concentration response (CR).
}
\details{
CR requires signature scores to have already been computed for a nullset.
randomdata() can generate a nullset, and this function can compute signature
scores for it by setting dataset = nullset and nullset = NULL. Pathway
scores are written to disk in output/signature_score_summary/. CR results
are written to disk in output/signature/conc_resp_summary/.
}
