% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureCatalogLoader.R
\name{signatureCatalogLoader}
\alias{signatureCatalogLoader}
\title{Merge the up and down halves of the pathway data}
\usage{
signatureCatalogLoader(
  sigset = "pilot_small",
  sigcatalog = "signatureDB_master_catalog 2020-01-31"
)
}
\arguments{
\item{sigset}{Name of the signature set.}

\item{sigcatlog}{Nmae of the catalog file}
}
\value{
the trimmed signature table
}
\description{
Merge the up and down halves of the pathway data
}
