% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureClassHM.R
\name{signatureClassHM}
\alias{signatureClassHM}
\title{Build lane plots by chemical list and signature class, across the datasets}
\usage{
signatureClassHM(
  to.file = F,
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  pathset = "PathwaySet_20191107",
  method = "gsva",
  threshold = 0.5
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Build lane plots by chemical list and signature class, across the datasets
}
