% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureConcResp.R
\name{signatureConcResp}
\alias{signatureConcResp}
\title{Pathway Concentration Response (P-value)}
\usage{
signatureConcResp(
  dataset,
  sigset,
  sigcatalog = "../inst/extdata/signatureDB_master_catalog_2022-05-16.xlsx",
  method,
  bmr_scale = 1.349,
  mc.cores = 1,
  pval = 0.05,
  nlowconc = 2,
  aicc = F,
  dtxsid.exclude = NULL,
  minsigsize = 10,
  fitmodels = c("cnst", "hill", "poly1", "poly2", "pow", "exp2", "exp3", "exp4", "exp5"),
  signaturescoremat,
  sigdbgenelist = "../inst/extdata/signatureDB_genelists.RDS",
  CHEM_DICT
)
}
\arguments{
\item{dataset}{Name of the data set.}

\item{sigset}{Name of the signature set.}

\item{sigcatalog}{Name of the signature catalog file}

\item{method}{Pathway scoring method in c("fc", "gsva", "gsea")}

\item{bmr_scale}{bmr scaling factor. Default = 1.349}

\item{mc.cores}{Number of cores to parallelize with.}

\item{pval}{Desired cutoff p-value.}

\item{nlowconc}{Only include the lowest nlowconc concentrations for each chemical}

\item{aicc}{aicc = T uses corrected AIC to choose winning method; otherwise regular AIC}

\item{dtxsid.exclude}{dtxsids to exclude, default NULL}

\item{minsigsize}{Minimum allowed signature size. Sample/signature combinations
with less than this number of non-missing l2fc's will be discarded.}

\item{fitmodels}{Vector of model names to use. Probably should include "cnst".}

\item{signaturescoremat}{dataframe returned by the upstream signatureScoreMerge function}

\item{sigdbgenelist}{full path to signature DB gene list file; default is repo version}

\item{CHEM_DICT}{Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}
}
\value{
dataframe with signature CR output.
}
\description{
Performs signature concentration response using p-value based cutoffs.
}
\details{
dataset should have already been scored using signatureScore and signatureScoreMerge
and the given sigset and method. This function prepares signatureScore output
for CR processing, calls tcplfit2::concRespCore, and formats the output.
}
