% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureHeatmap.R
\name{signatureHeatmap}
\alias{signatureHeatmap}
\title{Create heatmaps of the htpp data frames}
\usage{
signatureHeatmap(
  to.file = F,
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  sigset = "pilot_tiny",
  sigcatalog = "signatureDB_master_catalog 2020-01-31",
  method = "mygsea",
  conthits = T,
  nametag = "conthits",
  pval = 0.05
)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}

\item{dataset}{The set of data to be included}
}
\description{
Create heatmaps of the htpp data frames
}
