% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signaturePOD.R
\name{signaturePOD}
\alias{signaturePOD}
\title{Calculate PODs at the signature level}
\usage{
signaturePOD(
  do.load = F,
  sigset = "screen_large",
  dataset = "PFAS_U2OS",
  method = "fc",
  bmr_scale = 1.349,
  hccut = 0.95
)
}
\arguments{
\item{sigset}{Name of signature set.}

\item{dataset}{Name of data set.}

\item{method}{Pathway scoring method in c("fc", "gsva", "mygsea")}

\item{bmr_scale}{bmr scaling factor. Default = 1.349}

\item{hccut}{Remove rows with hitcall less than this value}

\item{do.laod}{If TRUE, load the input data into memory}
}
\description{
Calculate PODs at the signature level
}
