% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signaturePOD.R
\name{signaturePOD}
\alias{signaturePOD}
\title{Build lane plots by chemical list and signature class, across the datasets}
\usage{
signaturePOD(sigset, dataset, method, hit.threshold = 0.5)
}
\arguments{
\item{to.file}{If TRUE, write plots to a file}
}
\description{
Build lane plots by chemical list and signature class, across the datasets
}
