% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureRank.R
\name{signatureRank}
\alias{signatureRank}
\title{Get the signature ranks for chemicals}
\usage{
signatureRank(
  to.file = F,
  dataset = "DMEM_6hr_pilot_normal_pe_1",
  sigset = "pilot_large",
  sigcatalog = "signatureDB_master_catalog 2020-01-31",
  method = "mygsea"
)
}
\description{
Get the signature ranks for chemicals
}
