% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureScore.R
\name{signatureScore}
\alias{signatureScore}
\title{Signature Score}
\usage{
signatureScore(
  FCMAT2,
  CHEM_DICT,
  sigset,
  sigcatalog,
  dataset,
  method,
  mc.cores = 1,
  minsigsize = 10
)
}
\arguments{
\item{FCMAT2}{Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{CHEM_DICT}{Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}

\item{sigset}{Name of signature set.}

\item{dataset}{Name of data set.}

\item{method}{Signature scoring method in c("fc", "gsva", "mygsea")}

\item{mc.cores}{Number of cores to use.}

\item{minsigsize}{Minimum allowed signature size BEFORE accounting for
missing values.}
}
\value{
No output.
}
\description{
Computes and saves signature scores.
}
\details{
signatureScore is a driver for various scoring methods. The three that are
currently available are "gsva", "mygsea", "fc", and "mygsea_norank" (a version
of mygsea that uses fold changes instead of ranks as weights). Deprecated
methods include the Fisher method and gsvae (gsva with empirical cdfs).
Beware running out of memory on large runs with gsva, Linux, and many cores.
Signature size is counted according to number of genes in the signature that are
also in the column names of FCMAT2. However, each method performs a more
rigorous size count internally that accounts for missing values and adds this
to the output. This minsigsize is enforced when running signatureConcResp_pval.
}
