% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureScoreCoreGSEA.R
\name{signatureScoreCoreGSEA}
\alias{signatureScoreCoreGSEA}
\title{Signature Score Core - GSEA}
\usage{
signatureScoreCoreGSEA(
  sk.list,
  normfactor = 7500,
  sigset,
  fcmat,
  chem_dict,
  signature_data,
  mc.cores = 1,
  normalization = T,
  useranks = T
)
}
\arguments{
\item{sk.list}{Sample keys to use; should correspond to fcmat rownames.}

\item{normfactor}{= proceed on first 1/normfactor of GSEA data}

\item{sigset}{Name of signature set.}

\item{fcmat}{Expects FCMAT2. Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{chem_dict}{Expects CHEM_DICT object. Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}

\item{signature_data}{Named list of gene name vectors. Each element is one
signature, defined by the genes it contains.}

\item{mc.cores}{Number of cores to use.}

\item{normalization}{normalization = T normalizes final scores.}

\item{useranks}{useranks = T uses score ranks for weighting; otherwise,
fold changes are used for weights.}
}
\value{
Dataframe with one row per chemical/conc/signature combination. Columns
  are: sample_id, dtxsid, casrn, name, time, conc, sigset, signature, size, signature_score
}
\description{
Computes signature scores for gsea.
}
\details{
This function is a parallelized wrapper for gsea, which does the actual
scoring. gsea method uses ranks and normalizationby default. Normalization divides final
scores by difference between max and min score. Without normalization,
scores from individual samples have no impact on each other.
}
