% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureScoreCoreGSVA.R
\name{signatureScoreCoreGSVA}
\alias{signatureScoreCoreGSVA}
\title{Signature Score Core - GSVA}
\usage{
signatureScoreCoreGSVA(
  sk.list,
  sigset = "FILTERED",
  dataset,
  fcmat,
  chem_dict,
  signature_data,
  mc.cores = 1
)
}
\arguments{
\item{sk.list}{Sample keys to use; should correspond to fcmat rownames.}

\item{sigset}{Name of signature set.}

\item{dataset}{Name of data set.}

\item{fcmat}{Sample by gene matrix of log2(fold change)'s. Rownames are
sample keys and colnames are genes.}

\item{chem_dict}{Dataframe with one row per sample key and seven columns:
sample_key, sample_id, conc, time, casrn, name, dtxsid.}

\item{signature_data}{Named ist of gene name vectors. Each element is one
signature, defined by the genes it contains.}

\item{mc.cores}{Number of cores to use. Parallelization is performed
by gsva itself.}
}
\value{
No output.
}
\description{
Computes GSVA signature scores.
}
\details{
This function is a wrapper for GSVA with Gaussian cdf kernels. signaturescoremat
output is saved directly to disk.
}
