% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureScoreMerge.R
\name{signatureScoreMerge}
\alias{signatureScoreMerge}
\title{Merge the up and down halves of the pathway data}
\usage{
signatureScoreMerge(
  sigset = "pilot_small",
  sigcatalog = "signatureDB_master_catalog 2020-01-31",
  dataset = "DMEM_6hr_pilot_normal_pe_1_RAND1000",
  method = "mygsea",
  nullset = "DMEM_6hr_pilot_normal_pe_1_RAND1000"
)
}
\arguments{
\item{sigset}{Name of the signature set.}

\item{dataset}{Name of the data set.}

\item{method}{Pathway scoring method in c("fc", "gsva", "mygsea")}

\item{nullset}{Name of the null data set.}

\item{sigcatlog}{Nmae of the catalog file}
}
\value{
nothing
}
\description{
Merge the up and down halves of the pathway data
}
