% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureScoreMerge.R
\name{signatureScoreMerge}
\alias{signatureScoreMerge}
\title{Merge the up and down halves of the directional pathway data}
\usage{
signatureScoreMerge(
  sigset,
  sigcatalog = "../inst/extdata/signatureDB_master_catalog_2022-05-16.xlsx",
  method,
  signaturescoremat,
  sigdbgenelist = "../inst/extdata/signatureDB_genelists.RDS"
)
}
\arguments{
\item{sigset}{Name of the signature set.}

\item{sigcatalog}{Name of the catalog file}

\item{method}{Pathway scoring method in c("fc", "gsva", "gsea")}

\item{signaturescoremat}{dataframe returned by the upstream signatureScore function}

\item{sigdbgenelist}{full path to signature DB gene list file; default is repo version}
}
\value{
signaturescoremat dataframe
}
\description{
Merge the up and down halves of the directional pathway data
}
