% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxT.R
\name{TxT}
\alias{TxT}
\title{Calculate several statistics on a 2 x 2 matrix}
\usage{
TxT(tp, fp, fn, tn, do.p = TRUE, rowname = NA)
}
\arguments{
\item{tp}{number of true positives}

\item{fp}{number of false positives}

\item{fn}{number of false negatives}

\item{tn}{number of true negatives}

\item{do.p}{if TRUE, calculate an exact p-value}

\item{rowname}{if not NA, add a column to the output with this rowname}
}
\value{
Returns a list of the results with the following columns: TP (true positives), FP (false positives), FN (false negatives),
TN (true negatives), Sens (sensitivity), Spec (specificity), BA (Balanced Accuracy), Acrcy (Accuracy), RelRsk Relative Risk,
OR Odds Ratio, CI.OR.LWR (lower confidence interval of the Odds Ratio), CI.OR.UPR (upper confidence interval of the Odds Ratio),
PPV (Positive Predictive Value), NPV (Negative Predictive Value), F1 (2TP/(2TP+FP+FN)), p.value (Chi-squared p-value)
}
\description{
Calculate several statistics on a 2 x 2 matrix
}
